summary-methods {timeDate}	R Documentation

Summary Method

Description

Summarizes details of a 'timeDate' object.

Usage

## S3 method for class 'timeDate'
summary(object, ...)

Arguments

object
an object of class timeDate.

...
arguments passed to other methods.

Value

returns a summary report of the details of a timeDate object. This includes the starting and end date, the number of dates the format and the financial center in use.

Note

This S3 method will become in the future a S4 method

Examples

## summary -
   tC = timeCalendar()
   summary(tC)

[Package timeDate version 3043.102 Index]