Easter {timeDate}	R Documentation

Date of Easter

Description

Returns the date of Easter.

Usage

Easter(year = getRmetricsOptions("currentYear"), shift = 0)

Arguments

year
an integer value or integer vector for the year(s).

shift
an integer value, the number of days shifted from the Easter date. Negative integers are allowed.

Details

Holidays may have two origins, ecclesiastical and public/federal. The ecclesiastical calendars of Christian churches are based on cycles of moveable and immoveable feasts. Christmas, December 25th, is the principal immoveable feast. Easter is the principal moveable feast, and dates of most other moveable feasts are determined with respect to Easter.

The date of Easter is evaluated by a complex procedure whose detailed explanation goes beyond this description. The reason that the calculation is so complicate is, because the date of Easter is linked to (an inaccurate version of) the Hebrew calendar. But nevertheless a short answer to the question "When is Easter?" is the following: Easter Sunday is the first Sunday after the first full moon after vernal equinox. For the long answer we refer to Toendering (1998).

The algorithm computes the date of Easter based on the algorithm of Oudin (1940). It is valid for any Gregorian Calendar year.

Value

the date of Easter as an object of class "timeDate"

Note

Doesn't have options to compute Eastern Orthodox Easter dates.

See Also

BoxingDay, etc., for descriptions of the individual holiday functions,

listHolidays for a list (character vector) of all holidays,

holiday alternative to calling directly individual holiday functions (takes one or more holiday functions as argument),

holidayLONDON, holidayNERC, holidayNYSE, holidayTSX, holidayZURICH for holidays at major financial centers.

Examples

## Easter
# current year
Easter()
   
## From 2001 to 2010:
Easter(2001:2010)

[Package timeDate version 4052.112 Index]