dayOfWeek {timeDate}	R Documentation

Day of the week

Description

Returns the days of the week of the data in a "timeDate" object.

Usage

dayOfWeek(x)

Arguments

x
an object of class "timeDate".

Value

a character vector giving the days of the week corresponding to the elements of x. The names are in English, abbreviated to three letters.

See Also

dayOfYear

Examples

  
## timeCalendar
tC <- timeCalendar(2022)
       
## the days of the year
dayOfWeek(tC)

[Package timeDate version 4052.112 Index]