dayOfYear {timeDate}	R Documentation

Day of the year

Description

Returns the days of the year of the data in a "timeDate" object.

Usage

dayOfYear(x)

Arguments

x
an object of class "timeDate".

Value

vector of integers representing the number of days since the beginning of the year. For January, 1st it is one.

See Also

dayOfWeek

Examples

  
## timeCalendar
tC <- timeCalendar(2022)
       
## the days of the year
dayOfYear(tC)

[Package timeDate version 4052.112 Index]