isWeekday {timeDate}	R Documentation

Weekdays and Weekends

Description

Tests if a date is a weekday or not.

Usage

isWeekday(x, wday = 1:5)
isWeekend(x, wday = 1:5)

Arguments

x
an object of class timeDate.

wday
Specify which days should be considered as weekdays. By default from Mondays to Fridays.

Value

the functions return logical vectors indicating if a date is a weekday, or a weekend day.

Examples

## Dates in April, currentYear:
   currentYear = getRmetricsOptions("currentYear")
   tS = timeSequence(
      from = paste(currentYear, "-03-01", sep = ""),
      to = paste(currentYear, "-04-30", sep = ""))
   tS

## Subset of Weekends:
   isWeekend(tS)
   tS[isWeekend(tS)]

[Package timeDate version 3043.102 Index]