nDay {timeDate}	R Documentation

n-th n-day dates

Description

Computes the date for the n-th or last occurrence of an n-day in year/month.

Usage

timeNthNdayInMonth(charvec, nday = 1, nth = 1, format = "%Y-%m-%d",
    zone = "", FinCenter = "")

timeLastNdayInMonth(charvec, nday = 1, format = "%Y-%m-%d",
    zone = "", FinCenter = "")

Arguments

charvec
a character vector or object from a class representing time, such as "timeDate", "POSIXlt", etc.

nday
an integer vector with entries ranging from 0 (Sunday) to 6 (Saturday).

nth
an integer vector numbering the n-th occurence.

format
the format specification of the input character vector.

zone
the time zone or financial center where the data were recorded.

FinCenter
a character with the location of the financial center named as “continent/city”.

Details

timeNthNdayInMonth returns the nth occurrence of a n-day (nth = 1,...,5) in year, month.

timeLastNdayInMonth returns the last nday in year, month.

Value

an object of class "timeDate"

See Also

trunc.timeDate,

timeFirstDayInMonth, timeLastDayInMonth, timeFirstDayInQuarter, timeLastDayInQuarter,

timeNdayOnOrAfter, timeNdayOnOrBefore

Examples

## What date is the second Monday in April 2004?
timeNthNdayInMonth("2004-04-01", 1, 2)

## What date has the last Tuesday in May, 1996?
timeLastNdayInMonth("1996-05-01", 2)

[Package timeDate version 4052.112 Index]