earlyCloseNYSE {timeDate}	R Documentation

Early closings of the New York Stock exchange

Description

Get dates of early closings of the New York Stock exchange (NYSE).

Usage

earlyCloseNYSE(year)

Arguments

year
a vector of integers representing years (4 digits).

Details

earlyCloseNYSE gives the dates and times when NYSE was closed early. Some of these closing are scheduled (e.g. at 1pm on the day before or after a holiday), others are unscheduled.

The information is incomplete, particularly after 2011. For those dates the values are computed using explicitly declared rules or, if not available, ones derived from recent years.

Value

a "timeDate" object containing the dates (with closing times) of early closings

Note

The function is somewhat experimental but the type of the result will not change.

Author(s)

Georgi N. Boshnakov

References

https://archive.fo/XecDq

See Also

holidayNYSE for a list of NYSE holidays

Examples

earlyCloseNYSE(1990)

earlyCloseNYSE(2022:2024) # early closings
holidayNYSE(2022:2024)    # holidays
## early closings & holidays combined
c(earlyCloseNYSE(2022:2024), holidayNYSE(2022:2024))

[Package timeDate version 4052.112 Index]