Public and Ecclesiastical Holidays

A collection and description of functions and methods dealing with holiday dates in the G7 countries and Switzerland.

Usage

Septuagesima(year = getRmetricsOptions("currentYear"))
Quinquagesima(year = getRmetricsOptions("currentYear"))
AshWednesday(year = getRmetricsOptions("currentYear"))
PalmSunday(year = getRmetricsOptions("currentYear"))
GoodFriday(year = getRmetricsOptions("currentYear"))
EasterSunday(year = getRmetricsOptions("currentYear"))
EasterMonday(year = getRmetricsOptions("currentYear"))
RogationSunday(year = getRmetricsOptions("currentYear"))
Ascension(year = getRmetricsOptions("currentYear"))
Pentecost(year = getRmetricsOptions("currentYear"))
PentecostMonday(year = getRmetricsOptions("currentYear"))
TrinitySunday(year = getRmetricsOptions("currentYear"))
CorpusChristi(year = getRmetricsOptions("currentYear"))
ChristTheKing(year = getRmetricsOptions("currentYear"))
Advent1st(year = getRmetricsOptions("currentYear"))
Advent2nd(year = getRmetricsOptions("currentYear"))
Advent3rd(year = getRmetricsOptions("currentYear"))
Advent4th(year = getRmetricsOptions("currentYear"))
ChristmasEve(year = getRmetricsOptions("currentYear"))
ChristmasDay(year = getRmetricsOptions("currentYear"))
BoxingDay(year = getRmetricsOptions("currentYear"))
NewYearsDay(year = getRmetricsOptions("currentYear"))
SolemnityOfMary(year = getRmetricsOptions("currentYear"))
Epiphany(year = getRmetricsOptions("currentYear"))
PresentationOfLord(year = getRmetricsOptions("currentYear"))
Annunciation(year = getRmetricsOptions("currentYear"))
TransfigurationOfLord(year = getRmetricsOptions("currentYear"))
AssumptionOfMary(year = getRmetricsOptions("currentYear"))
BirthOfVirginMary(year = getRmetricsOptions("currentYear"))
CelebrationOfHolyCross(year = getRmetricsOptions("currentYear"))
MassOfArchangels(year = getRmetricsOptions("currentYear"))
AllSaints(year = getRmetricsOptions("currentYear"))
AllSouls(year = getRmetricsOptions("currentYear"))
LaborDay(year = getRmetricsOptions("currentYear"))
CHBerchtoldsDay(year = getRmetricsOptions("currentYear"))
CHSechselaeuten(year = getRmetricsOptions("currentYear"))
CHAscension(year = getRmetricsOptions("currentYear"))
CHConfederationDay(year = getRmetricsOptions("currentYear"))
CHKnabenschiessen(year = getRmetricsOptions("currentYear"))

GBMayDay(year = getRmetricsOptions("currentYear"))
GBEarlyMayBankHoliday(year = getRmetricsOptions("currentYear"))

GBBankHoliday(year = getRmetricsOptions("currentYear")) # see note in details section
GBSpringBankHoliday(year = getRmetricsOptions("currentYear"))

GBSummerBankHoliday(year = getRmetricsOptions("currentYear"))
GBMilleniumDay(year = getRmetricsOptions("currentYear"))
DEAscension(year = getRmetricsOptions("currentYear"))
DECorpusChristi(year = getRmetricsOptions("currentYear"))
DEGermanUnity(year = getRmetricsOptions("currentYear"))
DEChristmasEve(year = getRmetricsOptions("currentYear"))
DENewYearsEve(year = getRmetricsOptions("currentYear"))
FRFetDeLaVictoire1945(year = getRmetricsOptions("currentYear"))
FRAscension(year = getRmetricsOptions("currentYear"))
FRBastilleDay(year = getRmetricsOptions("currentYear"))
FRAssumptionVirginMary(year = getRmetricsOptions("currentYear"))
FRAllSaints(year = getRmetricsOptions("currentYear"))
FRArmisticeDay(year = getRmetricsOptions("currentYear"))
ITEpiphany(year = getRmetricsOptions("currentYear"))
ITLiberationDay(year = getRmetricsOptions("currentYear"))
ITAssumptionOfVirginMary(year = getRmetricsOptions("currentYear"))
ITAllSaints(year = getRmetricsOptions("currentYear"))
ITStAmrose(year = getRmetricsOptions("currentYear"))
ITImmaculateConception(year = getRmetricsOptions("currentYear"))
USDecorationMemorialDay(year = getRmetricsOptions("currentYear"))
USPresidentsDay(year = getRmetricsOptions("currentYear"))
USNewYearsDay(year = getRmetricsOptions("currentYear"))
USInaugurationDay(year = getRmetricsOptions("currentYear"))
USMLKingsBirthday(year = getRmetricsOptions("currentYear"))
USLincolnsBirthday(year = getRmetricsOptions("currentYear"))
USWashingtonsBirthday(year = getRmetricsOptions("currentYear"))
USMemorialDay(year = getRmetricsOptions("currentYear"))
USIndependenceDay(year = getRmetricsOptions("currentYear"))
USLaborDay(year = getRmetricsOptions("currentYear"))
USColumbusDay(year = getRmetricsOptions("currentYear"))
USElectionDay(year = getRmetricsOptions("currentYear"))
USVeteransDay(year = getRmetricsOptions("currentYear"))
USThanksgivingDay(year = getRmetricsOptions("currentYear"))
USChristmasDay(year = getRmetricsOptions("currentYear"))
USCPulaskisBirthday(year = getRmetricsOptions("currentYear"))
USGoodFriday(year = getRmetricsOptions("currentYear"))
USJuneteenthNationalIndependenceDay(year = getRmetricsOptions("currentYear"))
CAVictoriaDay(year = getRmetricsOptions("currentYear"))
CACanadaDay(year = getRmetricsOptions("currentYear"))
CACivicProvincialHoliday(year = getRmetricsOptions("currentYear"))
CALabourDay(year = getRmetricsOptions("currentYear"))
CAThanksgivingDay(year = getRmetricsOptions("currentYear"))
CaRemembranceDay(year = getRmetricsOptions("currentYear"))
JPVernalEquinox (year = getRmetricsOptions("currentYear"))
JPNewYearsDay(year = getRmetricsOptions("currentYear"))
JPGantan(year = getRmetricsOptions("currentYear"))
JPBankHolidayJan2(year = getRmetricsOptions("currentYear"))
JPBankHolidayJan3(year = getRmetricsOptions("currentYear"))
JPComingOfAgeDay(year = getRmetricsOptions("currentYear"))
JPSeijinNoHi(year = getRmetricsOptions("currentYear"))
JPNatFoundationDay(year = getRmetricsOptions("currentYear"))
JPKenkokuKinenNoHi(year = getRmetricsOptions("currentYear"))
JPGreeneryDay(year = getRmetricsOptions("currentYear"))
JPMidoriNoHi(year = getRmetricsOptions("currentYear"))
JPConstitutionDay(year = getRmetricsOptions("currentYear"))
JPKenpouKinenBi(year = getRmetricsOptions("currentYear"))
JPNationHoliday(year = getRmetricsOptions("currentYear"))
JPKokuminNoKyujitu(year = getRmetricsOptions("currentYear"))
JPChildrensDay(year = getRmetricsOptions("currentYear"))
JPKodomoNoHi(year = getRmetricsOptions("currentYear"))
JPMarineDay(year = getRmetricsOptions("currentYear"))
JPUmiNoHi(year = getRmetricsOptions("currentYear"))
JPRespectForTheAgedDay(year = getRmetricsOptions("currentYear"))
JPKeirouNOhi(year = getRmetricsOptions("currentYear"))
JPAutumnalEquinox(year = getRmetricsOptions("currentYear"))
JPShuubunNoHi(year = getRmetricsOptions("currentYear"))
JPHealthandSportsDay(year = getRmetricsOptions("currentYear"))
JPTaiikuNoHi(year = getRmetricsOptions("currentYear"))
JPNationalCultureDay(year = getRmetricsOptions("currentYear"))
JPBunkaNoHi(year = getRmetricsOptions("currentYear"))
JPThanksgivingDay(year = getRmetricsOptions("currentYear"))
JPKinrouKanshaNoHi(year = getRmetricsOptions("currentYear"))
JPEmperorsBirthday(year = getRmetricsOptions("currentYear"))
JPTennouTanjyouBi(year = getRmetricsOptions("currentYear"))
JPBankHolidayDec31(year = getRmetricsOptions("currentYear"))

Arguments

year

an integer value or vector of year numbers including the century. These are integers of the form CCYY, e.g. 2000.

Value

The function listHolidays returns a character vector with the names of the supported holidays.

The holiday functions return an ISO-8601 formatted 'timeDate' of the requested holiday.

Details

The deprecated GBMayDay and GBBankHoliday are still available but strongly discouraged. Instead, use GBEarlyMayBankHoliday and GBSpringBankHoliday, respectively.

Examples

## listHolidays -
   listHolidays()
#>   [1] "Advent1st"                          
#>   [2] "Advent2nd"                          
#>   [3] "Advent3rd"                          
#>   [4] "Advent4th"                          
#>   [5] "AllSaints"                          
#>   [6] "AllSouls"                           
#>   [7] "Annunciation"                       
#>   [8] "Ascension"                          
#>   [9] "AshWednesday"                       
#>  [10] "AssumptionOfMary"                   
#>  [11] "BirthOfVirginMary"                  
#>  [12] "BoxingDay"                          
#>  [13] "CACanadaDay"                        
#>  [14] "CACivicProvincialHoliday"           
#>  [15] "CALabourDay"                        
#>  [16] "CAThanksgivingDay"                  
#>  [17] "CAVictoriaDay"                      
#>  [18] "CHAscension"                        
#>  [19] "CHBerchtoldsDay"                    
#>  [20] "CHConfederationDay"                 
#>  [21] "CHKnabenschiessen"                  
#>  [22] "CHSechselaeuten"                    
#>  [23] "CaRemembranceDay"                   
#>  [24] "CelebrationOfHolyCross"             
#>  [25] "ChristTheKing"                      
#>  [26] "ChristmasDay"                       
#>  [27] "ChristmasEve"                       
#>  [28] "CorpusChristi"                      
#>  [29] "DEAscension"                        
#>  [30] "DEChristmasEve"                     
#>  [31] "DECorpusChristi"                    
#>  [32] "DEGermanUnity"                      
#>  [33] "DENewYearsEve"                      
#>  [34] "Easter"                             
#>  [35] "EasterMonday"                       
#>  [36] "EasterSunday"                       
#>  [37] "Epiphany"                           
#>  [38] "FRAllSaints"                        
#>  [39] "FRArmisticeDay"                     
#>  [40] "FRAscension"                        
#>  [41] "FRAssumptionVirginMary"             
#>  [42] "FRBastilleDay"                      
#>  [43] "FRFetDeLaVictoire1945"              
#>  [44] "GBEarlyMayBankHoliday"              
#>  [45] "GBMilleniumDay"                     
#>  [46] "GBSpringBankHoliday"                
#>  [47] "GBSummerBankHoliday"                
#>  [48] "GoodFriday"                         
#>  [49] "ITAllSaints"                        
#>  [50] "ITAssumptionOfVirginMary"           
#>  [51] "ITEpiphany"                         
#>  [52] "ITImmaculateConception"             
#>  [53] "ITLiberationDay"                    
#>  [54] "ITStAmrose"                         
#>  [55] "JPAutumnalEquinox"                  
#>  [56] "JPBankHolidayDec31"                 
#>  [57] "JPBankHolidayJan2"                  
#>  [58] "JPBankHolidayJan3"                  
#>  [59] "JPBunkaNoHi"                        
#>  [60] "JPChildrensDay"                     
#>  [61] "JPComingOfAgeDay"                   
#>  [62] "JPConstitutionDay"                  
#>  [63] "JPEmperorsBirthday"                 
#>  [64] "JPGantan"                           
#>  [65] "JPGreeneryDay"                      
#>  [66] "JPHealthandSportsDay"               
#>  [67] "JPKeirouNOhi"                       
#>  [68] "JPKenkokuKinenNoHi"                 
#>  [69] "JPKenpouKinenBi"                    
#>  [70] "JPKinrouKanshaNoHi"                 
#>  [71] "JPKodomoNoHi"                       
#>  [72] "JPKokuminNoKyujitu"                 
#>  [73] "JPMarineDay"                        
#>  [74] "JPMidoriNoHi"                       
#>  [75] "JPNatFoundationDay"                 
#>  [76] "JPNationHoliday"                    
#>  [77] "JPNationalCultureDay"               
#>  [78] "JPNewYearsDay"                      
#>  [79] "JPRespectForTheAgedDay"             
#>  [80] "JPSeijinNoHi"                       
#>  [81] "JPShuubunNoHi"                      
#>  [82] "JPTaiikuNoHi"                       
#>  [83] "JPTennouTanjyouBi"                  
#>  [84] "JPThanksgivingDay"                  
#>  [85] "JPUmiNoHi"                          
#>  [86] "LaborDay"                           
#>  [87] "MassOfArchangels"                   
#>  [88] "NewYearsDay"                        
#>  [89] "PalmSunday"                         
#>  [90] "Pentecost"                          
#>  [91] "PentecostMonday"                    
#>  [92] "PresentationOfLord"                 
#>  [93] "Quinquagesima"                      
#>  [94] "RogationSunday"                     
#>  [95] "Septuagesima"                       
#>  [96] "SolemnityOfMary"                    
#>  [97] "TransfigurationOfLord"              
#>  [98] "TrinitySunday"                      
#>  [99] "USCPulaskisBirthday"                
#> [100] "USChristmasDay"                     
#> [101] "USColumbusDay"                      
#> [102] "USDecorationMemorialDay"            
#> [103] "USElectionDay"                      
#> [104] "USGoodFriday"                       
#> [105] "USInaugurationDay"                  
#> [106] "USIndependenceDay"                  
#> [107] "USJuneteenthNationalIndependenceDay"
#> [108] "USLaborDay"                         
#> [109] "USLincolnsBirthday"                 
#> [110] "USMLKingsBirthday"                  
#> [111] "USMemorialDay"                      
#> [112] "USNewYearsDay"                      
#> [113] "USPresidentsDay"                    
#> [114] "USThanksgivingDay"                  
#> [115] "USVeteransDay"                      
#> [116] "USWashingtonsBirthday"              

## CHSechselaeuten -
   # Sechselaeuten a half Day Bank Holiday in Switzerland
   CHSechselaeuten(2000:2010)
#> GMT
#>  [1] [2000-04-17] [2001-04-16] [2002-04-15] [2003-04-21] [2004-04-19]
#>  [6] [2005-04-18] [2006-04-17] [2007-04-16] [2008-04-21] [2009-04-20]
#> [11] [2010-04-19]
   CHSechselaeuten(getRmetricsOptions("currentYear"))
#> GMT
#> [1] [2022-04-18]

## German Unification Day:
   DEGermanUnity(getRmetricsOptions("currentYear"))
#> GMT
#> [1] [2022-10-03]