listHolidays {timeDate}	R Documentation

List of holidays

Description

Returns a list of holidays supported by package "timeDate".

Usage

listHolidays(pattern = ".*") 

Arguments

pattern
a character string containing a regular expression.

Details

Gives a character vector containing the names of supported holidays matching pattern. The default is to return all holidays.

The list is sorted alphabetically. It is changed from time to time. So, the use of character indexing (possibly representing patterns) on the returned list is strongly recommended.

Value

a character vector, sorted in alphabetical order

See Also

BoxingDay, etc., for descriptions of the individual holiday functions,

holiday alternative to calling directly individual holiday functions (takes one or more holiday functions as argument),

Easter,

holidayLONDON, holidayNERC, holidayNYSE, holidayTSX, holidayZURICH for holidays at major financial centers.

Examples

## Local Swiss Holidays:
listHolidays("CH")

listHolidays("Easter")
listHolidays("NewYear")

## All Holidays
listHolidays()

[Package timeDate version 4052.112 Index]