holidayNERC {timeDate}	R Documentation

NERC holiday calendar

Description

Returns a holiday calendar for NERC, the North American Reliability Council.

Usage

holidayNERC(year = getRmetricsOptions("currentYear"), FinCenter = "Eastern")

Arguments

year
an integer value or vector of years, formatted as YYYY.

FinCenter
a character value, the name of the financial center to use.

Value

an object of class "timeDate"

Author(s)

Joe W. Byers

References

http://www.nerc.com/~oc/offpeaks.html

See Also

holidayLONDON, holidayNYSE, holidayTSX, holidayZURICH for holidays at other major financial centers.

BoxingDay, etc., for descriptions of the individual holiday functions,

listHolidays for a list (character vector) of all holidays,

holiday alternative to calling directly individual holiday functions (takes one or more holiday functions as argument),

Easter

Examples

holidayNERC()
holidayNERC(2008:2010)

[Package timeDate version 4052.112 Index]