holidayTSX {timeDate}	R Documentation

TSX holiday calendar

Description

Returns a holiday calendar for the Toronto Stock Exchange.

Usage

holidayTSX(year = getRmetricsOptions("currentYear"))

Arguments

year
an integer value or vector of years, formatted as YYYY.

Value

an object of class "timeDate"

See Also

holidayLONDON, holidayNERC, holidayNYSE, holidayZURICH for holidays at other major financial centers,

BoxingDay, etc., for descriptions of the individual holiday functions,

listHolidays for a list (character vector) of all holidays,

holiday alternative to calling directly individual holiday functions (takes one or more holiday functions as argument),

Easter

Examples

holidayTSX()
holidayTSX(2008:2010)

[Package timeDate version 4052.112 Index]