coerceToOther {timeDate}	R Documentation

Coercion from 'timeDate' to other classes

Description

Coerce and transform objects of class "timeDate".

Usage

## S3 method for class 'timeDate'
as.character(x, ...)

## S3 method for class 'timeDate'
as.double(x, 
    units = c("auto", "secs", "mins", "hours", "days", "weeks"), ...)

## S3 method for class 'timeDate'
as.data.frame(x, ...)

## S3 method for class 'timeDate'
as.POSIXct(x, tz = "", ...)

## S3 method for class 'timeDate'
as.POSIXlt(x, tz = "", ...)

## S3 method for class 'timeDate'
as.Date(x, method = c("trunc", "round", "next"), ...)

Arguments

x
an object of class "timeDate".

units
a character string denoting the date/time units in which the results are desired.

tz
inputs the time zone to POSIX objects, i.e. the time zone, zone, or financial center string, FinCenter, as used by "timeDate" objects.

method
a character string denoting the method how to determine the dates.

...
arguments passed to other methods.

Value

an object from the designated target class

See Also

timeDate and as.timeDate for creation of and conversion to "timeDate" objects

Examples

## timeDate
tC = timeCalendar()

## convert 'timeDate' to a character vector
as.character(tC)

[Package timeDate version 4052.112 Index]