names-methods {timeDate}	R Documentation

The names of a 'timeDate' object

Description

Functions to get or set the names of a "timeDate" object.

Usage

## S4 method for signature 'timeDate'
names(x)
## S4 replacement method for signature 'timeDate'
names(x) <- value

Arguments

x
an object of class "timeDate".

value
a character vector of up to the same length as x, or NULL.

Examples

td <- timeCalendar()
td
names(td) <- LETTERS[seq_along(td)]
td

[Package timeDate version 4052.112 Index]