myFinCenter {timeDate}	R Documentation

myFinCenter variable

Description

A character string with the name of my financial center.

Note

Can be modified by the user to his/her own or any other financial center. The default is "GMT". To list all supported financial centers use the function listFinCenter.

See Also

listFinCenter

Examples

## myFinCenter - the global setting currently used
getRmetricsOptions("myFinCenter")
   
## change to another financial center
# setRmetricsOptions(myFinCenter = "Zurich")
   
## Do not care about DST
# setRmetricsOptions(myFinCenter = "GMT") 

[Package timeDate version 4052.112 Index]