Daylight Saving Time Rules

Functions for about 400 cities and regions which return daylight saving time rules and time zone offsets.

Details

As a selection of these functions:

Adelaide Algiers Amsterdam Anchorage Andorra Athens Auckland Bahrain Bangkok Beirut Belfast Belgrade Berlin Bogota Bratislava Brisbane Brussels Bucharest Budapest BuenosAires Cairo Calcutta Caracas Casablanca Cayman Chicago Copenhagen Darwin Denver Detroit Dubai Dublin Eastern Edmonton Frankfurt Helsinki HongKong Honolulu Indianapolis Istanbul Jakarta Jerusalem Johannesburg Kiev KualaLumpur Kuwait Lagos Lisbon Ljubljana London LosAngeles Luxembourg Madrid Manila Melbourne MexicoCity Monaco Montreal Moscow Nairobi Nassau NewYork Nicosia Oslo Pacific Paris Perth Prague Riga Riyadh Rome Seoul Shanghai Singapore Sofia Stockholm Sydney Taipei Tallinn Tehran Tokyo Tunis Vaduz Vancouver Vienna Vilnius Warsaw Winnipeg Zagreb Zurich, ...

Note

There are currently two synonyms available "Pacific" for Los Angeles and "Eastern" for New York.

Specific time zones (AST, CET, CST, EET, EST, MST and PST) are also available.

Note we leave the space in all double named cities like New York or Hong Kong and use an underscore for it.

All the entries are retrieved from the tzdata library which is available under GNU GPL licence.

Examples

## DST Rules for Zurich:
   head(Zurich())
#>                Zurich offSet isdst TimeZone     numeric
#> 1 1901-12-14 20:45:52   3600     0      CET -2147397248
#> 2 1941-05-05 00:00:00   7200     1     CEST  -904435200
#> 3 1941-10-06 00:00:00   3600     0      CET  -891129600
#> 4 1942-05-04 00:00:00   7200     1     CEST  -872985600
#> 5 1942-10-05 00:00:00   3600     0      CET  -859680000
#> 6 1981-03-29 01:00:00   7200     1     CEST   354675600
   tail(Zurich())
#>                  Zurich offSet isdst TimeZone    numeric
#> 114 2035-03-25 01:00:00   7200     1     CEST 2058397200
#> 115 2035-10-28 01:00:00   3600     0      CET 2077146000
#> 116 2036-03-30 01:00:00   7200     1     CEST 2090451600
#> 117 2036-10-26 01:00:00   3600     0      CET 2108595600
#> 118 2037-03-29 01:00:00   7200     1     CEST 2121901200
#> 119 2037-10-25 01:00:00   3600     0      CET 2140045200

## list all available centers
listFinCenter()
#>   [1] "Africa/Abidjan"                 "Africa/Accra"                  
#>   [3] "Africa/Addis_Ababa"             "Africa/Algiers"                
#>   [5] "Africa/Asmara"                  "Africa/Bamako"                 
#>   [7] "Africa/Bangui"                  "Africa/Banjul"                 
#>   [9] "Africa/Bissau"                  "Africa/Blantyre"               
#>  [11] "Africa/Brazzaville"             "Africa/Bujumbura"              
#>  [13] "Africa/Cairo"                   "Africa/Casablanca"             
#>  [15] "Africa/Ceuta"                   "Africa/Conakry"                
#>  [17] "Africa/Dakar"                   "Africa/Dar_es_Salaam"          
#>  [19] "Africa/Djibouti"                "Africa/Douala"                 
#>  [21] "Africa/El_Aaiun"                "Africa/Freetown"               
#>  [23] "Africa/Gaborone"                "Africa/Harare"                 
#>  [25] "Africa/Johannesburg"            "Africa/Kampala"                
#>  [27] "Africa/Khartoum"                "Africa/Kigali"                 
#>  [29] "Africa/Kinshasa"                "Africa/Lagos"                  
#>  [31] "Africa/Libreville"              "Africa/Lome"                   
#>  [33] "Africa/Luanda"                  "Africa/Lubumbashi"             
#>  [35] "Africa/Lusaka"                  "Africa/Malabo"                 
#>  [37] "Africa/Maputo"                  "Africa/Maseru"                 
#>  [39] "Africa/Mbabane"                 "Africa/Mogadishu"              
#>  [41] "Africa/Monrovia"                "Africa/Nairobi"                
#>  [43] "Africa/Ndjamena"                "Africa/Niamey"                 
#>  [45] "Africa/Nouakchott"              "Africa/Ouagadougou"            
#>  [47] "Africa/Porto-Novo"              "Africa/Sao_Tome"               
#>  [49] "Africa/Tripoli"                 "Africa/Tunis"                  
#>  [51] "Africa/Windhoek"                "America/Adak"                  
#>  [53] "America/Anchorage"              "America/Anguilla"              
#>  [55] "America/Antigua"                "America/Araguaina"             
#>  [57] "America/Argentina/Buenos_Aires" "America/Argentina/Catamarca"   
#>  [59] "America/Argentina/Cordoba"      "America/Argentina/Jujuy"       
#>  [61] "America/Argentina/La_Rioja"     "America/Argentina/Mendoza"     
#>  [63] "America/Argentina/Rio_Gallegos" "America/Argentina/San_Juan"    
#>  [65] "America/Argentina/Tucuman"      "America/Argentina/Ushuaia"     
#>  [67] "America/Aruba"                  "America/Asuncion"              
#>  [69] "America/Atikokan"               "America/Bahia"                 
#>  [71] "America/Barbados"               "America/Belem"                 
#>  [73] "America/Belize"                 "America/Blanc-Sablon"          
#>  [75] "America/Boa_Vista"              "America/Bogota"                
#>  [77] "America/Boise"                  "America/Cambridge_Bay"         
#>  [79] "America/Campo_Grande"           "America/Cancun"                
#>  [81] "America/Caracas"                "America/Cayenne"               
#>  [83] "America/Cayman"                 "America/Chicago"               
#>  [85] "America/Chihuahua"              "America/Costa_Rica"            
#>  [87] "America/Cuiaba"                 "America/Curacao"               
#>  [89] "America/Danmarkshavn"           "America/Dawson"                
#>  [91] "America/Dawson_Creek"           "America/Denver"                
#>  [93] "America/Detroit"                "America/Dominica"              
#>  [95] "America/Edmonton"               "America/Eirunepe"              
#>  [97] "America/El_Salvador"            "America/Fortaleza"             
#>  [99] "America/Glace_Bay"              "America/Godthab"               
#> [101] "America/Goose_Bay"              "America/Grand_Turk"            
#> [103] "America/Grenada"                "America/Guadeloupe"            
#> [105] "America/Guatemala"              "America/Guayaquil"             
#> [107] "America/Guyana"                 "America/Halifax"               
#> [109] "America/Havana"                 "America/Hermosillo"            
#> [111] "America/Indiana/Indianapolis"   "America/Indiana/Knox"          
#> [113] "America/Indiana/Marengo"        "America/Indiana/Petersburg"    
#> [115] "America/Indiana/Tell_City"      "America/Indiana/Vevay"         
#> [117] "America/Indiana/Vincennes"      "America/Indiana/Winamac"       
#> [119] "America/Inuvik"                 "America/Iqaluit"               
#> [121] "America/Jamaica"                "America/Juneau"                
#> [123] "America/Kentucky/Louisville"    "America/Kentucky/Monticello"   
#> [125] "America/La_Paz"                 "America/Lima"                  
#> [127] "America/Los_Angeles"            "America/Maceio"                
#> [129] "America/Managua"                "America/Manaus"                
#> [131] "America/Marigot"                "America/Martinique"            
#> [133] "America/Mazatlan"               "America/Menominee"             
#> [135] "America/Merida"                 "America/Mexico_City"           
#> [137] "America/Miquelon"               "America/Moncton"               
#> [139] "America/Monterrey"              "America/Montevideo"            
#> [141] "America/Montreal"               "America/Montserrat"            
#> [143] "America/Nassau"                 "America/New_York"              
#> [145] "America/Nipigon"                "America/Nome"                  
#> [147] "America/Noronha"                "America/North_Dakota/Center"   
#> [149] "America/North_Dakota/New_Salem" "America/Panama"                
#> [151] "America/Pangnirtung"            "America/Paramaribo"            
#> [153] "America/Phoenix"                "America/Port-au-Prince"        
#> [155] "America/Port_of_Spain"          "America/Porto_Velho"           
#> [157] "America/Puerto_Rico"            "America/Rainy_River"           
#> [159] "America/Rankin_Inlet"           "America/Recife"                
#> [161] "America/Regina"                 "America/Resolute"              
#> [163] "America/Rio_Branco"             "America/Santiago"              
#> [165] "America/Santo_Domingo"          "America/Sao_Paulo"             
#> [167] "America/Scoresbysund"           "America/Shiprock"              
#> [169] "America/St_Barthelemy"          "America/St_Johns"              
#> [171] "America/St_Kitts"               "America/St_Lucia"              
#> [173] "America/St_Thomas"              "America/St_Vincent"            
#> [175] "America/Swift_Current"          "America/Tegucigalpa"           
#> [177] "America/Thule"                  "America/Thunder_Bay"           
#> [179] "America/Tijuana"                "America/Toronto"               
#> [181] "America/Tortola"                "America/Vancouver"             
#> [183] "America/Whitehorse"             "America/Winnipeg"              
#> [185] "America/Yakutat"                "America/Yellowknife"           
#> [187] "Antarctica/Casey"               "Antarctica/Davis"              
#> [189] "Antarctica/DumontDUrville"      "Antarctica/Mawson"             
#> [191] "Antarctica/McMurdo"             "Antarctica/Palmer"             
#> [193] "Antarctica/Rothera"             "Antarctica/South_Pole"         
#> [195] "Antarctica/Syowa"               "Antarctica/Vostok"             
#> [197] "Arctic/Longyearbyen"            "Asia/Aden"                     
#> [199] "Asia/Almaty"                    "Asia/Amman"                    
#> [201] "Asia/Anadyr"                    "Asia/Aqtau"                    
#> [203] "Asia/Aqtobe"                    "Asia/Ashgabat"                 
#> [205] "Asia/Baghdad"                   "Asia/Bahrain"                  
#> [207] "Asia/Baku"                      "Asia/Bangkok"                  
#> [209] "Asia/Beirut"                    "Asia/Bishkek"                  
#> [211] "Asia/Brunei"                    "Asia/Calcutta"                 
#> [213] "Asia/Choibalsan"                "Asia/Chongqing"                
#> [215] "Asia/Colombo"                   "Asia/Damascus"                 
#> [217] "Asia/Dhaka"                     "Asia/Dili"                     
#> [219] "Asia/Dubai"                     "Asia/Dushanbe"                 
#> [221] "Asia/Gaza"                      "Asia/Harbin"                   
#> [223] "Asia/Hong_Kong"                 "Asia/Hovd"                     
#> [225] "Asia/Irkutsk"                   "Asia/Jakarta"                  
#> [227] "Asia/Jayapura"                  "Asia/Jerusalem"                
#> [229] "Asia/Kabul"                     "Asia/Kamchatka"                
#> [231] "Asia/Karachi"                   "Asia/Kashgar"                  
#> [233] "Asia/Katmandu"                  "Asia/Krasnoyarsk"              
#> [235] "Asia/Kuala_Lumpur"              "Asia/Kuching"                  
#> [237] "Asia/Kuwait"                    "Asia/Macau"                    
#> [239] "Asia/Magadan"                   "Asia/Makassar"                 
#> [241] "Asia/Manila"                    "Asia/Muscat"                   
#> [243] "Asia/Nicosia"                   "Asia/Novosibirsk"              
#> [245] "Asia/Omsk"                      "Asia/Oral"                     
#> [247] "Asia/Phnom_Penh"                "Asia/Pontianak"                
#> [249] "Asia/Pyongyang"                 "Asia/Qatar"                    
#> [251] "Asia/Qyzylorda"                 "Asia/Rangoon"                  
#> [253] "Asia/Riyadh"                    "Asia/Saigon"                   
#> [255] "Asia/Sakhalin"                  "Asia/Samarkand"                
#> [257] "Asia/Seoul"                     "Asia/Shanghai"                 
#> [259] "Asia/Singapore"                 "Asia/Taipei"                   
#> [261] "Asia/Tashkent"                  "Asia/Tbilisi"                  
#> [263] "Asia/Tehran"                    "Asia/Thimphu"                  
#> [265] "Asia/Tokyo"                     "Asia/Ulaanbaatar"              
#> [267] "Asia/Urumqi"                    "Asia/Vientiane"                
#> [269] "Asia/Vladivostok"               "Asia/Yakutsk"                  
#> [271] "Asia/Yekaterinburg"             "Asia/Yerevan"                  
#> [273] "Atlantic/Azores"                "Atlantic/Bermuda"              
#> [275] "Atlantic/Canary"                "Atlantic/Cape_Verde"           
#> [277] "Atlantic/Faroe"                 "Atlantic/Madeira"              
#> [279] "Atlantic/Reykjavik"             "Atlantic/South_Georgia"        
#> [281] "Atlantic/St_Helena"             "Atlantic/Stanley"              
#> [283] "Australia/Adelaide"             "Australia/Brisbane"            
#> [285] "Australia/Broken_Hill"          "Australia/Currie"              
#> [287] "Australia/Darwin"               "Australia/Eucla"               
#> [289] "Australia/Hobart"               "Australia/Lindeman"            
#> [291] "Australia/Lord_Howe"            "Australia/Melbourne"           
#> [293] "Australia/Perth"                "Australia/Sydney"              
#> [295] "Europe/Amsterdam"               "Europe/Andorra"                
#> [297] "Europe/Athens"                  "Europe/Belgrade"               
#> [299] "Europe/Berlin"                  "Europe/Bratislava"             
#> [301] "Europe/Brussels"                "Europe/Bucharest"              
#> [303] "Europe/Budapest"                "Europe/Chisinau"               
#> [305] "Europe/Copenhagen"              "Europe/Dublin"                 
#> [307] "Europe/Gibraltar"               "Europe/Guernsey"               
#> [309] "Europe/Helsinki"                "Europe/Isle_of_Man"            
#> [311] "Europe/Istanbul"                "Europe/Jersey"                 
#> [313] "Europe/Kaliningrad"             "Europe/Kiev"                   
#> [315] "Europe/Lisbon"                  "Europe/Ljubljana"              
#> [317] "Europe/London"                  "Europe/Luxembourg"             
#> [319] "Europe/Madrid"                  "Europe/Malta"                  
#> [321] "Europe/Mariehamn"               "Europe/Minsk"                  
#> [323] "Europe/Monaco"                  "Europe/Moscow"                 
#> [325] "Europe/Oslo"                    "Europe/Paris"                  
#> [327] "Europe/Podgorica"               "Europe/Prague"                 
#> [329] "Europe/Riga"                    "Europe/Rome"                   
#> [331] "Europe/Samara"                  "Europe/San_Marino"             
#> [333] "Europe/Sarajevo"                "Europe/Simferopol"             
#> [335] "Europe/Skopje"                  "Europe/Sofia"                  
#> [337] "Europe/Stockholm"               "Europe/Tallinn"                
#> [339] "Europe/Tirane"                  "Europe/Uzhgorod"               
#> [341] "Europe/Vaduz"                   "Europe/Vatican"                
#> [343] "Europe/Vienna"                  "Europe/Vilnius"                
#> [345] "Europe/Volgograd"               "Europe/Warsaw"                 
#> [347] "Europe/Zagreb"                  "Europe/Zaporozhye"             
#> [349] "Europe/Zurich"                  "Indian/Antananarivo"           
#> [351] "Indian/Chagos"                  "Indian/Christmas"              
#> [353] "Indian/Cocos"                   "Indian/Comoro"                 
#> [355] "Indian/Kerguelen"               "Indian/Mahe"                   
#> [357] "Indian/Maldives"                "Indian/Mauritius"              
#> [359] "Indian/Mayotte"                 "Indian/Reunion"                
#> [361] "Pacific/Apia"                   "Pacific/Auckland"              
#> [363] "Pacific/Chatham"                "Pacific/Efate"                 
#> [365] "Pacific/Enderbury"              "Pacific/Fakaofo"               
#> [367] "Pacific/Fiji"                   "Pacific/Funafuti"              
#> [369] "Pacific/Galapagos"              "Pacific/Gambier"               
#> [371] "Pacific/Guadalcanal"            "Pacific/Guam"                  
#> [373] "Pacific/Honolulu"               "Pacific/Johnston"              
#> [375] "Pacific/Kiritimati"             "Pacific/Kosrae"                
#> [377] "Pacific/Kwajalein"              "Pacific/Majuro"                
#> [379] "Pacific/Marquesas"              "Pacific/Midway"                
#> [381] "Pacific/Nauru"                  "Pacific/Niue"                  
#> [383] "Pacific/Norfolk"                "Pacific/Noumea"                
#> [385] "Pacific/Pago_Pago"              "Pacific/Palau"                 
#> [387] "Pacific/Pitcairn"               "Pacific/Ponape"                
#> [389] "Pacific/Port_Moresby"           "Pacific/Rarotonga"             
#> [391] "Pacific/Saipan"                 "Pacific/Tahiti"                
#> [393] "Pacific/Tarawa"                 "Pacific/Tongatapu"             
#> [395] "Pacific/Truk"                   "Pacific/Wake"                  
#> [397] "Pacific/Wallis"