Sys.timeDate {timeDate}	R Documentation

System time as 'timeDate' object

Description

Returns the system time as an object of class "timeDate".

Usage

Sys.timeDate(FinCenter = "")

Arguments

FinCenter
a character string with the location of the financial center named as "continent/city".

Value

a "timeDate" object

Examples

## Not run: 
## direct
Sys.timeDate()

## Local Time in Zurich
Sys.timeDate(FinCenter = "Zurich")  

## transformed from "POSIX(c)t" with timeDate()
timeDate(Sys.time()) 
 
## Local Time in Zurich
timeDate(Sys.time(), FinCenter = "Zurich")  

## End(Not run)   

[Package timeDate version 4052.112 Index]