finCenter {timeDate}	R Documentation

Financial Center of a timeDate object

Description

Get or set the financial center of a "timeDate" object.

Usage

## S4 method for signature 'timeDate'
finCenter(x)
## S4 replacement method for signature 'timeDate'
finCenter(x) <- value

Arguments

x
a timeSeries object.

value
a character with the location of the financial center named as "continent/city".

Details

"timeDate" objects store the time in the GMT time zone. The financial center specifies a location whose local time is to be used to format the object, e.g., for printing.

finCenter gives the financial center associated with a ‘timeDate’ object. The assignment form changes it to the specified value. Both functions are S4 generics. This page describes the methods defined in package ‘timeDate’.

See Also

listFinCenter

Examples

date <- timeDate("2008-01-01")
finCenter(date) <- "GMT"
date
format(date)

finCenter(date) <- "Zurich"
date
format(date)

[Package timeDate version 4052.112 Index]