whichFormat {timeDate}	R Documentation

Format recognition

Description

Tries to recognize the date/time format.

Usage

whichFormat(charvec, silent = FALSE)  

Arguments

charvec
a character string or vector of dates and times.

silent
a logical flag. Should a warning be printed if the format cannot be recognized?

Value

a format string

See Also

midnightStandard

Examples

whichFormat("2007-12-31 00:00:00")
whichFormat("2007-12-31 24:00")

[Package timeDate version 4052.112 Index]