diff {timeDate}	R Documentation

Lagged 'timeDate' Differences

Description

Returns suitably lagged and iterated differences.

Usage

## S3 method for class 'timeDate'
diff(x, lag = 1, differences = 1, ...)  

Arguments

x
an object of class timeDate.

lag
an integer indicating which lag to use.

differences
an integer indicating the order of the difference.

...
arguments passed to other methods.

Value

For the function, diff.timeDate, if x is a vector of length n and differences=1, then the computed result is equal to the successive differences x[(1+lag):n] - x[1:(n-lag)]. If difference is larger than one this algorithm is applied recursively to x. Note that the returned value is a vector which is shorter than x.

Examples

## Create Character Vectors:
   dts = c("1989-09-28", "2001-01-15", "2004-08-30", "1990-02-09")
   dts
   tms = c(  "23:12:55",   "10:34:02",   "08:30:00",   "11:18:23")
   tms
   
## timeDate - 
   GMT = timeDate(dts, zone = "GMT", FinCenter = "GMT") + 24*3600
   GMT
   ZUR = timeDate(dts, zone = "GMT", FinCenter = "Europe/Zurich")
   ZUR

## diff - 
   # Suitably Lagged and Iterated Differences:
   diff(GMT)
   diff(GMT, lag = 2)
   diff(GMT, lag = 1, diff = 2)

[Package timeDate version 3043.102 Index]