start {timeDate}	R Documentation

Terminal Times and Range

Description

Extracts the time the first or last observation was taken, or computes the range.

Usage

## S3 method for class 'timeDate'
start(x, ...)

## S3 method for class 'timeDate'
end(x, ...)

## S3 method for class 'timeDate'
min(..., na.rm = FALSE)

## S3 method for class 'timeDate'
max(..., na.rm = FALSE)

## S3 method for class 'timeDate'
range(..., na.rm = FALSE)

Arguments

x
an object of class timeDate.

...
[start][end] -
not used,
[min][max] - 'timeDates' objects.

na.rm
not used.

Details

Note, the series will be time ordered before the start or end time stamps are extracted. Sorting is done in the way that the first observation appears in time before the last observation.

Value

returns an object of class timeDate.

Examples

  
## timeCalendar - 
   # Random Calendar Dates:
   tR = sample(timeCalendar())
   sort(tR)
   tR
    
## start | end  -
   start(tR)
   end(tR)
   
## The First and Last Time Stamp:
   tR[1]
   tR[length(tR)]
   rev(tR)[1]
   
## The Range:
   c(start(tR), end(tR))
   range(tR)

[Package timeDate version 3043.102 Index]