Align a 'timeDate' object to regular date/time stamps

Aligns a "timeDate" object to regular date/time stamps.

Usage

# S4 method for timeDate
align(x, by = "1d", offset = "0s")

alignDaily(x, include.weekends=FALSE)
alignMonthly(x, include.weekends=FALSE)
alignQuarterly(x, include.weekends=FALSE)

Arguments

x

an object of class "timeDate".

by

a character string formed from an integer length and a period identifier. Valid values are "w", "d", "h", "m", "s", for weeks, days, hours, minutes and seconds. For example a bi-weekly period is expressed as "2w".

offset

a character string to set an offset formed from an integer length and a period identifier in the same way as for argument by.

include.weekends

logical value indicating whether weekends should be included.

Details

The functions alignDaily, alignMonthly, alignMonthly are simple to use functions which generate end-of-day, end-of-month, and end-of quarter "timeDate" objects. Weekends are excluded by default. Optionally they can be added setting the argument include.weekends = TRUE.

Value

an object of class "timeDate"

Examples

## align -

# Align Bi-Weekly with a 3 Days Offset:
(tC <- timeCalendar())
#> GMT
#>  [1] [2022-01-01] [2022-02-01] [2022-03-01] [2022-04-01] [2022-05-01]
#>  [6] [2022-06-01] [2022-07-01] [2022-08-01] [2022-09-01] [2022-10-01]
#> [11] [2022-11-01] [2022-12-01]
align(tC, by = "2w", offset = "3d")
#> GMT
#>  [1] [2022-01-04] [2022-01-18] [2022-02-01] [2022-02-15] [2022-03-01]
#>  [6] [2022-03-15] [2022-03-29] [2022-04-12] [2022-04-26] [2022-05-10]
#> [11] [2022-05-24] [2022-06-07] [2022-06-21] [2022-07-05] [2022-07-19]
#> [16] [2022-08-02] [2022-08-16] [2022-08-30] [2022-09-13] [2022-09-27]
#> [21] [2022-10-11] [2022-10-25] [2022-11-08] [2022-11-22]

## alignDaily -

# Simple to use Functions:
alignDaily(tC)
#> GMT
#>   [1] [2022-01-03] [2022-01-04] [2022-01-05] [2022-01-06] [2022-01-07]
#>   [6] [2022-01-10] [2022-01-11] [2022-01-12] [2022-01-13] [2022-01-14]
#>  [11] [2022-01-17] [2022-01-18] [2022-01-19] [2022-01-20] [2022-01-21]
#>  [16] [2022-01-24] [2022-01-25] [2022-01-26] [2022-01-27] [2022-01-28]
#>  [21] [2022-01-31] [2022-02-01] [2022-02-02] [2022-02-03] [2022-02-04]
#>  [26] [2022-02-07] [2022-02-08] [2022-02-09] [2022-02-10] [2022-02-11]
#>  [31] [2022-02-14] [2022-02-15] [2022-02-16] [2022-02-17] [2022-02-18]
#>  [36] [2022-02-21] [2022-02-22] [2022-02-23] [2022-02-24] [2022-02-25]
#>  [41] [2022-02-28] [2022-03-01] [2022-03-02] [2022-03-03] [2022-03-04]
#>  [46] [2022-03-07] [2022-03-08] [2022-03-09] [2022-03-10] [2022-03-11]
#>  [51] [2022-03-14] [2022-03-15] [2022-03-16] [2022-03-17] [2022-03-18]
#>  [56] [2022-03-21] [2022-03-22] [2022-03-23] [2022-03-24] [2022-03-25]
#>  [61] [2022-03-28] [2022-03-29] [2022-03-30] [2022-03-31] [2022-04-01]
#>  [66] [2022-04-04] [2022-04-05] [2022-04-06] [2022-04-07] [2022-04-08]
#>  [71] [2022-04-11] [2022-04-12] [2022-04-13] [2022-04-14] [2022-04-15]
#>  [76] [2022-04-18] [2022-04-19] [2022-04-20] [2022-04-21] [2022-04-22]
#>  [81] [2022-04-25] [2022-04-26] [2022-04-27] [2022-04-28] [2022-04-29]
#>  [86] [2022-05-02] [2022-05-03] [2022-05-04] [2022-05-05] [2022-05-06]
#>  [91] [2022-05-09] [2022-05-10] [2022-05-11] [2022-05-12] [2022-05-13]
#>  [96] [2022-05-16] [2022-05-17] [2022-05-18] [2022-05-19] [2022-05-20]
#> ...
#>  [ reached getRmetricsOption('max.print') | getOption('max.print') -- omitted 139 rows ]]
alignDaily(tC, include.weekends=TRUE)
#> GMT
#>   [1] [2022-01-01] [2022-01-02] [2022-01-03] [2022-01-04] [2022-01-05]
#>   [6] [2022-01-06] [2022-01-07] [2022-01-08] [2022-01-09] [2022-01-10]
#>  [11] [2022-01-11] [2022-01-12] [2022-01-13] [2022-01-14] [2022-01-15]
#>  [16] [2022-01-16] [2022-01-17] [2022-01-18] [2022-01-19] [2022-01-20]
#>  [21] [2022-01-21] [2022-01-22] [2022-01-23] [2022-01-24] [2022-01-25]
#>  [26] [2022-01-26] [2022-01-27] [2022-01-28] [2022-01-29] [2022-01-30]
#>  [31] [2022-01-31] [2022-02-01] [2022-02-02] [2022-02-03] [2022-02-04]
#>  [36] [2022-02-05] [2022-02-06] [2022-02-07] [2022-02-08] [2022-02-09]
#>  [41] [2022-02-10] [2022-02-11] [2022-02-12] [2022-02-13] [2022-02-14]
#>  [46] [2022-02-15] [2022-02-16] [2022-02-17] [2022-02-18] [2022-02-19]
#>  [51] [2022-02-20] [2022-02-21] [2022-02-22] [2022-02-23] [2022-02-24]
#>  [56] [2022-02-25] [2022-02-26] [2022-02-27] [2022-02-28] [2022-03-01]
#>  [61] [2022-03-02] [2022-03-03] [2022-03-04] [2022-03-05] [2022-03-06]
#>  [66] [2022-03-07] [2022-03-08] [2022-03-09] [2022-03-10] [2022-03-11]
#>  [71] [2022-03-12] [2022-03-13] [2022-03-14] [2022-03-15] [2022-03-16]
#>  [76] [2022-03-17] [2022-03-18] [2022-03-19] [2022-03-20] [2022-03-21]
#>  [81] [2022-03-22] [2022-03-23] [2022-03-24] [2022-03-25] [2022-03-26]
#>  [86] [2022-03-27] [2022-03-28] [2022-03-29] [2022-03-30] [2022-03-31]
#>  [91] [2022-04-01] [2022-04-02] [2022-04-03] [2022-04-04] [2022-04-05]
#>  [96] [2022-04-06] [2022-04-07] [2022-04-08] [2022-04-09] [2022-04-10]
#> ...
#>  [ reached getRmetricsOption('max.print') | getOption('max.print') -- omitted 235 rows ]]

# Align to end-of-month Dates:
alignMonthly(tC)
#> GMT
#>  [1] [2022-01-31] [2022-02-28] [2022-03-31] [2022-04-29] [2022-05-31]
#>  [6] [2022-06-30] [2022-07-30] [2022-08-31] [2022-09-30] [2022-10-31]
#> [11] [2022-11-30] [2022-12-30]