dayOfWeek {timeDate}	R Documentation

Day of the Week

Description

returns the day of the year from a 'timeDate' object.

Usage

dayOfWeek(x)

Arguments

x
an object of class timeDate.

Value

returns a three letter character string with the names in English of the day of the week,

Note

With version 2.7 the function has been renamed from getdayOfWeek.

See Also

dayOfYear

Examples

  
## timeCalendat -  
   tC = timeCalendar()
       
## The days of the Year:
   dayOfWeek(tC)
   
## Use Deprecated Function:
   getDayOfWeek <- dayOfWeek
   getDayOfWeek(tC)

[Package timeDate version 3043.102 Index]