.endpoints {timeDate}	R Documentation

Endpoints Indexes

Description

Returns endpoint indexes from a 'timeDate' object

Usage

.endpoints(x, on = c("months", "years", "quarters", "weeks", "days",
    "hours", "minutes", "seconds"), k=1)

Arguments

x
a timeDate object.

on
the periods endpoints to find as a character string. Select from: "months", "years", "quarters", "weeks", "days", "hours", "minutes", "seconds".

k
along every k-th element.

Details

endpoints returns an integer vector corresponding to the last observation in each period specified by on, with a zero added to the beginning of the vector, and the index of the last observation in x at the end.

Value

returns an intege vector of endpoints beginning with 0 and ending with the a value equal to the length of the x argument.

Author(s)

Jeff Ryan, modofied by Diethelm Wuertz for timeDate objects.

Examples

## endpoints -

   # Weekly Endpoints:
   .endpoints(timeCalendar(), on="w")

[Package timeDate version 3043.102 Index]