Check if dates are business or holidays

Tests if a date is a business day or not.

Usage

isBizday(x, holidays = holidayNYSE(), wday = 1:5)
isHoliday(x, holidays = holidayNYSE(), wday = 1:5)

Arguments

x

an object of class "timeDate".

holidays

holiday dates from a holiday calendar. An object of class "timeDate".

wday

Specify which days should be considered as weekdays. By default from Mondays to Fridays.

Details

Returns a logical vector of the same length as x indicating if a date is a business day, or a holiday, respectively.

Value

a logical vector of the same length as x

Examples

## dates in April, current year
currentYear <- getRmetricsOptions("currentYear")
tS <- timeSequence(from = paste(currentYear, "-03-01", sep = ""),
                     to = paste(currentYear, "-04-30", sep = ""))
tS
#> GMT
#>  [1] [2025-03-01] [2025-03-02] [2025-03-03] [2025-03-04] [2025-03-05]
#>  [6] [2025-03-06] [2025-03-07] [2025-03-08] [2025-03-09] [2025-03-10]
#> [11] [2025-03-11] [2025-03-12] [2025-03-13] [2025-03-14] [2025-03-15]
#> [16] [2025-03-16] [2025-03-17] [2025-03-18] [2025-03-19] [2025-03-20]
#> [21] [2025-03-21] [2025-03-22] [2025-03-23] [2025-03-24] [2025-03-25]
#> [26] [2025-03-26] [2025-03-27] [2025-03-28] [2025-03-29] [2025-03-30]
#> [31] [2025-03-31] [2025-04-01] [2025-04-02] [2025-04-03] [2025-04-04]
#> [36] [2025-04-05] [2025-04-06] [2025-04-07] [2025-04-08] [2025-04-09]
#> [41] [2025-04-10] [2025-04-11] [2025-04-12] [2025-04-13] [2025-04-14]
#> [46] [2025-04-15] [2025-04-16] [2025-04-17] [2025-04-18] [2025-04-19]
#> [51] [2025-04-20] [2025-04-21] [2025-04-22] [2025-04-23] [2025-04-24]
#> [56] [2025-04-25] [2025-04-26] [2025-04-27] [2025-04-28] [2025-04-29]
#> [61] [2025-04-30]

## subset business days at NYSE
holidayNYSE()
#> NewYork
#>  [1] [2025-01-01] [2025-01-20] [2025-02-17] [2025-04-18] [2025-05-26]
#>  [6] [2025-06-19] [2025-07-04] [2025-09-01] [2025-11-27] [2025-12-25]
isBizday(tS, holidayNYSE())
#> 2025-03-01 2025-03-02 2025-03-03 2025-03-04 2025-03-05 2025-03-06 2025-03-07 
#>      FALSE      FALSE       TRUE       TRUE       TRUE       TRUE       TRUE 
#> 2025-03-08 2025-03-09 2025-03-10 2025-03-11 2025-03-12 2025-03-13 2025-03-14 
#>      FALSE      FALSE       TRUE       TRUE       TRUE       TRUE       TRUE 
#> 2025-03-15 2025-03-16 2025-03-17 2025-03-18 2025-03-19 2025-03-20 2025-03-21 
#>      FALSE      FALSE       TRUE       TRUE       TRUE       TRUE       TRUE 
#> 2025-03-22 2025-03-23 2025-03-24 2025-03-25 2025-03-26 2025-03-27 2025-03-28 
#>      FALSE      FALSE       TRUE       TRUE       TRUE       TRUE       TRUE 
#> 2025-03-29 2025-03-30 2025-03-31 2025-04-01 2025-04-02 2025-04-03 2025-04-04 
#>      FALSE      FALSE       TRUE       TRUE       TRUE       TRUE       TRUE 
#> 2025-04-05 2025-04-06 2025-04-07 2025-04-08 2025-04-09 2025-04-10 2025-04-11 
#>      FALSE      FALSE       TRUE       TRUE       TRUE       TRUE       TRUE 
#> 2025-04-12 2025-04-13 2025-04-14 2025-04-15 2025-04-16 2025-04-17 2025-04-18 
#>      FALSE      FALSE       TRUE       TRUE       TRUE       TRUE      FALSE 
#> 2025-04-19 2025-04-20 2025-04-21 2025-04-22 2025-04-23 2025-04-24 2025-04-25 
#>      FALSE      FALSE       TRUE       TRUE       TRUE       TRUE       TRUE 
#> 2025-04-26 2025-04-27 2025-04-28 2025-04-29 2025-04-30 
#>      FALSE      FALSE       TRUE       TRUE       TRUE 
tS[isBizday(tS, holidayNYSE())]
#> GMT
#>  [1] [2025-03-03] [2025-03-04] [2025-03-05] [2025-03-06] [2025-03-07]
#>  [6] [2025-03-10] [2025-03-11] [2025-03-12] [2025-03-13] [2025-03-14]
#> [11] [2025-03-17] [2025-03-18] [2025-03-19] [2025-03-20] [2025-03-21]
#> [16] [2025-03-24] [2025-03-25] [2025-03-26] [2025-03-27] [2025-03-28]
#> [21] [2025-03-31] [2025-04-01] [2025-04-02] [2025-04-03] [2025-04-04]
#> [26] [2025-04-07] [2025-04-08] [2025-04-09] [2025-04-10] [2025-04-11]
#> [31] [2025-04-14] [2025-04-15] [2025-04-16] [2025-04-17] [2025-04-21]
#> [36] [2025-04-22] [2025-04-23] [2025-04-24] [2025-04-25] [2025-04-28]
#> [41] [2025-04-29] [2025-04-30]