Check if dates are business or holidays

Tests if a date is a business day or not.

Usage

isBizday(x, holidays = holidayNYSE(), wday = 1:5)
isHoliday(x, holidays = holidayNYSE(), wday = 1:5)

Arguments

x

an object of class "timeDate".

holidays

holiday dates from a holiday calendar. An object of class "timeDate".

wday

Specify which days should be considered as weekdays. By default from Mondays to Fridays.

Details

Returns a logical vector of the same length as x indicating if a date is a business day, or a holiday, respectively.

Value

a logical vector of the same length as x

Examples

## Dates in April, currentYear:
   currentYear = getRmetricsOptions("currentYear")
   tS = timeSequence(
      from = paste(currentYear, "-03-01", sep = ""),
      to = paste(currentYear, "-04-30", sep = ""))
   tS
#> GMT
#>  [1] [2022-03-01] [2022-03-02] [2022-03-03] [2022-03-04] [2022-03-05]
#>  [6] [2022-03-06] [2022-03-07] [2022-03-08] [2022-03-09] [2022-03-10]
#> [11] [2022-03-11] [2022-03-12] [2022-03-13] [2022-03-14] [2022-03-15]
#> [16] [2022-03-16] [2022-03-17] [2022-03-18] [2022-03-19] [2022-03-20]
#> [21] [2022-03-21] [2022-03-22] [2022-03-23] [2022-03-24] [2022-03-25]
#> [26] [2022-03-26] [2022-03-27] [2022-03-28] [2022-03-29] [2022-03-30]
#> [31] [2022-03-31] [2022-04-01] [2022-04-02] [2022-04-03] [2022-04-04]
#> [36] [2022-04-05] [2022-04-06] [2022-04-07] [2022-04-08] [2022-04-09]
#> [41] [2022-04-10] [2022-04-11] [2022-04-12] [2022-04-13] [2022-04-14]
#> [46] [2022-04-15] [2022-04-16] [2022-04-17] [2022-04-18] [2022-04-19]
#> [51] [2022-04-20] [2022-04-21] [2022-04-22] [2022-04-23] [2022-04-24]
#> [56] [2022-04-25] [2022-04-26] [2022-04-27] [2022-04-28] [2022-04-29]
#> [61] [2022-04-30]

## Subset Business Days at NYSE:
   holidayNYSE()
#> NewYork
#> [1] [2022-01-17] [2022-02-21] [2022-04-15] [2022-05-30] [2022-06-20]
#> [6] [2022-07-04] [2022-09-05] [2022-11-24] [2022-12-26]
   isBizday(tS, holidayNYSE())
#> 2022-03-01 2022-03-02 2022-03-03 2022-03-04 2022-03-05 2022-03-06 2022-03-07 
#>       TRUE       TRUE       TRUE       TRUE      FALSE      FALSE       TRUE 
#> 2022-03-08 2022-03-09 2022-03-10 2022-03-11 2022-03-12 2022-03-13 2022-03-14 
#>       TRUE       TRUE       TRUE       TRUE      FALSE      FALSE       TRUE 
#> 2022-03-15 2022-03-16 2022-03-17 2022-03-18 2022-03-19 2022-03-20 2022-03-21 
#>       TRUE       TRUE       TRUE       TRUE      FALSE      FALSE       TRUE 
#> 2022-03-22 2022-03-23 2022-03-24 2022-03-25 2022-03-26 2022-03-27 2022-03-28 
#>       TRUE       TRUE       TRUE       TRUE      FALSE      FALSE       TRUE 
#> 2022-03-29 2022-03-30 2022-03-31 2022-04-01 2022-04-02 2022-04-03 2022-04-04 
#>       TRUE       TRUE       TRUE       TRUE      FALSE      FALSE       TRUE 
#> 2022-04-05 2022-04-06 2022-04-07 2022-04-08 2022-04-09 2022-04-10 2022-04-11 
#>       TRUE       TRUE       TRUE       TRUE      FALSE      FALSE       TRUE 
#> 2022-04-12 2022-04-13 2022-04-14 2022-04-15 2022-04-16 2022-04-17 2022-04-18 
#>       TRUE       TRUE       TRUE      FALSE      FALSE      FALSE       TRUE 
#> 2022-04-19 2022-04-20 2022-04-21 2022-04-22 2022-04-23 2022-04-24 2022-04-25 
#>       TRUE       TRUE       TRUE       TRUE      FALSE      FALSE       TRUE 
#> 2022-04-26 2022-04-27 2022-04-28 2022-04-29 2022-04-30 
#>       TRUE       TRUE       TRUE       TRUE      FALSE 
   tS[isBizday(tS, holidayNYSE())]
#> GMT
#>  [1] [2022-03-01] [2022-03-02] [2022-03-03] [2022-03-04] [2022-03-07]
#>  [6] [2022-03-08] [2022-03-09] [2022-03-10] [2022-03-11] [2022-03-14]
#> [11] [2022-03-15] [2022-03-16] [2022-03-17] [2022-03-18] [2022-03-21]
#> [16] [2022-03-22] [2022-03-23] [2022-03-24] [2022-03-25] [2022-03-28]
#> [21] [2022-03-29] [2022-03-30] [2022-03-31] [2022-04-01] [2022-04-04]
#> [26] [2022-04-05] [2022-04-06] [2022-04-07] [2022-04-08] [2022-04-11]
#> [31] [2022-04-12] [2022-04-13] [2022-04-14] [2022-04-18] [2022-04-19]
#> [36] [2022-04-20] [2022-04-21] [2022-04-22] [2022-04-25] [2022-04-26]
#> [41] [2022-04-27] [2022-04-28] [2022-04-29]