julian {timeDate}	R Documentation

Julian Counts and Calendar Atoms

Description

Returns Julian day counts, date/time atoms from a 'timeDate' object, and extracts month atoms from a 'timeDate' object.

Usage

## S4 method for signature 'timeDate'
julian(x, origin = timeDate("1970-01-01"), 
    units = c("auto", "secs", "mins", "hours", "days", "weeks"), 
    zone = NULL, FinCenter = NULL, ...)

## S4 method for signature 'timeDate'
atoms(x, ...)

## S4 method for signature 'timeDate'
months(x, abbreviate = NULL)

Arguments

x
an object of class timeDate.

origin
a length-one object inheriting from class "timeDate" setting the origin for the julian counter.

units
a character string denoting the date/time units in which the results are desired.

zone
the time zone or financial center where the data were recorded.

FinCenter
a character with the the location of the financial center named as "continent/city".

abbreviate
currently not used.

...
arguments passed to other methods.

Value

julian returns a timeDate object as a Julian count.

atoms and months extrac from a timeDate object the calendar atoms, i.e. the year, month, day, and optionally hour, minute and second.

Examples

## julian -
   tC = timeCalendar()
   julian(tC)[1:3]
   
## atoms -
   atoms(tC)
   
## months -
   months(tC)

[Package timeDate version 3043.102 Index]