listHolidays {timeDate}	R Documentation

List of Holidayss

Description

Returns the list of holidays.

Usage

listHolidays(pattern = ".*") 

Arguments

pattern
a pattern character string as required by the grep function.

Value

returns a list of holidays as a character vector.

Examples

## listHolidays -

   # All Holidays:
   listHolidays()
   
   # Local Swiss Holidays:
   listHolidays("CH") 

[Package timeDate version 3012.100 Index]