List of holidays

Returns a list of holidays.

Usage

listHolidays(pattern = ".*")

Arguments

pattern

a pattern character string as required by the grep function.

Details

Returns a character vector containing the names of supported holidays matching pattern. The default is to return all holidays.

Value

a character vector

Examples

## listHolidays -

   # All Holidays:
   listHolidays()
#>   [1] "Advent1st"                          
#>   [2] "Advent2nd"                          
#>   [3] "Advent3rd"                          
#>   [4] "Advent4th"                          
#>   [5] "AllSaints"                          
#>   [6] "AllSouls"                           
#>   [7] "Annunciation"                       
#>   [8] "Ascension"                          
#>   [9] "AshWednesday"                       
#>  [10] "AssumptionOfMary"                   
#>  [11] "BirthOfVirginMary"                  
#>  [12] "BoxingDay"                          
#>  [13] "CACanadaDay"                        
#>  [14] "CACivicProvincialHoliday"           
#>  [15] "CAFamilyDay"                        
#>  [16] "CALabourDay"                        
#>  [17] "CAThanksgivingDay"                  
#>  [18] "CAVictoriaDay"                      
#>  [19] "CHAscension"                        
#>  [20] "CHBerchtoldsDay"                    
#>  [21] "CHConfederationDay"                 
#>  [22] "CHKnabenschiessen"                  
#>  [23] "CHSechselaeuten"                    
#>  [24] "CaRemembranceDay"                   
#>  [25] "CelebrationOfHolyCross"             
#>  [26] "ChristTheKing"                      
#>  [27] "ChristmasDay"                       
#>  [28] "ChristmasEve"                       
#>  [29] "CorpusChristi"                      
#>  [30] "DEAscension"                        
#>  [31] "DEChristmasEve"                     
#>  [32] "DECorpusChristi"                    
#>  [33] "DEGermanUnity"                      
#>  [34] "DENewYearsEve"                      
#>  [35] "Easter"                             
#>  [36] "EasterMonday"                       
#>  [37] "EasterSunday"                       
#>  [38] "Epiphany"                           
#>  [39] "FRAllSaints"                        
#>  [40] "FRArmisticeDay"                     
#>  [41] "FRAscension"                        
#>  [42] "FRAssumptionVirginMary"             
#>  [43] "FRBastilleDay"                      
#>  [44] "FRFetDeLaVictoire1945"              
#>  [45] "GBEarlyMayBankHoliday"              
#>  [46] "GBSpringBankHoliday"                
#>  [47] "GBSummerBankHoliday"                
#>  [48] "GoodFriday"                         
#>  [49] "ITAllSaints"                        
#>  [50] "ITAssumptionOfVirginMary"           
#>  [51] "ITEpiphany"                         
#>  [52] "ITImmaculateConception"             
#>  [53] "ITLiberationDay"                    
#>  [54] "ITStAmrose"                         
#>  [55] "JPAutumnalEquinox"                  
#>  [56] "JPBankHolidayDec31"                 
#>  [57] "JPBankHolidayJan2"                  
#>  [58] "JPBankHolidayJan3"                  
#>  [59] "JPBunkaNoHi"                        
#>  [60] "JPChildrensDay"                     
#>  [61] "JPComingOfAgeDay"                   
#>  [62] "JPConstitutionDay"                  
#>  [63] "JPEmperorsBirthday"                 
#>  [64] "JPGantan"                           
#>  [65] "JPGreeneryDay"                      
#>  [66] "JPHealthandSportsDay"               
#>  [67] "JPKeirouNoHi"                       
#>  [68] "JPKenkokuKinenNoHi"                 
#>  [69] "JPKenpouKinenBi"                    
#>  [70] "JPKinrouKanshaNoHi"                 
#>  [71] "JPKodomoNoHi"                       
#>  [72] "JPKokuminNoKyujitu"                 
#>  [73] "JPMarineDay"                        
#>  [74] "JPMidoriNoHi"                       
#>  [75] "JPMountainDay"                      
#>  [76] "JPNatFoundationDay"                 
#>  [77] "JPNationHoliday"                    
#>  [78] "JPNationalCultureDay"               
#>  [79] "JPNewYearsDay"                      
#>  [80] "JPRespectForTheAgedDay"             
#>  [81] "JPSeijinNoHi"                       
#>  [82] "JPShuubunNoHi"                      
#>  [83] "JPTaiikuNoHi"                       
#>  [84] "JPTennouTanjyouBi"                  
#>  [85] "JPThanksgivingDay"                  
#>  [86] "JPUmiNoHi"                          
#>  [87] "JPVernalEquinox"                    
#>  [88] "LaborDay"                           
#>  [89] "MassOfArchangels"                   
#>  [90] "NewYearsDay"                        
#>  [91] "PalmSunday"                         
#>  [92] "Pentecost"                          
#>  [93] "PentecostMonday"                    
#>  [94] "PresentationOfLord"                 
#>  [95] "Quinquagesima"                      
#>  [96] "RogationSunday"                     
#>  [97] "Septuagesima"                       
#>  [98] "SolemnityOfMary"                    
#>  [99] "TransfigurationOfLord"              
#> [100] "TrinitySunday"                      
#> [101] "USCPulaskisBirthday"                
#> [102] "USChristmasDay"                     
#> [103] "USColumbusDay"                      
#> [104] "USDecorationMemorialDay"            
#> [105] "USElectionDay"                      
#> [106] "USGoodFriday"                       
#> [107] "USInaugurationDay"                  
#> [108] "USIndependenceDay"                  
#> [109] "USJuneteenthNationalIndependenceDay"
#> [110] "USLaborDay"                         
#> [111] "USLincolnsBirthday"                 
#> [112] "USMLKingsBirthday"                  
#> [113] "USMemorialDay"                      
#> [114] "USNewYearsDay"                      
#> [115] "USPresidentsDay"                    
#> [116] "USThanksgivingDay"                  
#> [117] "USVeteransDay"                      
#> [118] "USWashingtonsBirthday"              
#> [119] "specialHolidayGB"                   
   
   # Local Swiss Holidays:
   listHolidays("CH") 
#> [1] "CHAscension"        "CHBerchtoldsDay"    "CHConfederationDay"
#> [4] "CHKnabenschiessen"  "CHSechselaeuten"