holidayNERC {timeDate}	R Documentation

NERC Holiday Calendar

Description

Returns a holiday calendar for NERC, the North American Reliability Council.

Usage

holidayNERC(year = getRmetricsOptions("currentYear"), FinCenter = "Eastern")

Arguments

year
an integer value or vector of years, formatted as YYYY.

FinCenter
a character value, the name of the financial center to use.

Value

returns an object of class timeDate.

Author(s)

Joe W. Byers

References

http://www.nerc.com/~oc/offpeaks.html

Examples

## holidayNERC -
   holidayNERC()
   holidayNERC(2008:2010)

[Package timeDate version 3043.102 Index]