holidayNYSE {timeDate}	R Documentation

NYSE Holiday Calendar

Description

Returns a holiday calendar for the New York Stock Exchange.

Usage

holidayNYSE(year = getRmetricsOptions("currentYear"))

Arguments

year
an integer value or vector of years, formatted as YYYY.

Value

returns an object of class timeDate.

Examples

## holidayNYSE -
   holidayNYSE()
   holidayNYSE(2008:2010)

[Package timeDate version 3043.102 Index]