timeDateMathOps {timeDate}	R Documentation

timeDate Mathematical Operations

Description

Functions for mathematical and logical operations on 'timeDate' objects.

The functions are:

Ops,timeDate 	 Group 'Ops' generic functions for 'timeDate' objects,

+,timeDate 	 Performs arithmetic + operation on 'timeDate' objects,

-,timeDate 	 Performs arithmetic - operation on 'timeDate' objects.

Usage

## S4 method for signature 'timeDate,timeDate'
Ops(e1, e2)

Arguments

e1, e2
usually objects of class timeDate, in the case of addition and subtraction e2 may be of class numeric.

Value

Ops.timeDate
these are functions for mathematical operations. Group Ops are generic functions which manage mathematical operations.

+.timeDate
-.timeDate
The plus operator "+" performs arithmetic "+" operation on timeDate objects, and the minus operator "-" returns a difftime object if both arguments e1 and e2 are "timeDate" objects, or returns a "timeDate" object e2 seconds earlier than e1.

Examples

## Create Character Vectors:
   dts = c("1989-09-28", "2001-01-15", "2004-08-30", "1990-02-09")
   dts
   tms = c(  "23:12:55",   "10:34:02",   "08:30:00",   "11:18:23")
   tms
   
## "+/-" - 
   # Add One Day to a Given timeDate Object:
   GMT = timeDate(dts, zone = "GMT", FinCenter = "GMT")
   GMT
   ZUR = timeDate(dts, zone = "GMT", FinCenter = "Europe/Zurich")
   ZUR
   GMT + 24*3600
   ZUR[2] - ZUR[1] 

[Package timeDate version 3043.102 Index]