RmetricsOptions {timeDate}	R Documentation

Rmetrics Option Settings

Description

Allow the user to set and examine a variety of global options which affect the way in which Rmetrics functions computes and displays its results.

Usage

setRmetricsOptions(...)
getRmetricsOption(x, unset = "")

Arguments

unset
a character string holding the return value is x is not set.

x
a character string holding an option name.

...
any options can be defined, using name = value or by passing a list of such tagged values.

[Package timeDate version 3043.102 Index]