myFinCenter {timeDate}	R Documentation

myFinCenter Variable

Description

A character string with the name of my financial center.

Note

Can be modified by the user to his own or any other financial center. The default is "GMT". To list all supported financial center use the function listFinCenter.

See Also

listFinCenter

Examples

## myFinCenter - the global setting currently used:
   getRmetricsOptions("myFinCenter")
   
## Change to another Financiel Center:
   # setRmetricsOptions(myFinCenter = "Zurich")
   
## Do not take care about DST ...
   # setRmetricsOptions(myFinCenter = "GMT") 

[Package timeDate version 3043.102 Index]