blockStart {timeDate}	R Documentation

Equally sized 'timeDate' Blocks

Description

Creates start (end) dates for equally sized 'timeDate' blocks.

Usage

blockStart(x, block = 20)
blockEnd(x, block = 20)

Arguments

block
an integer value specifying the length in number of records for numerically sized blocks of dates.

x
an object of class timeDate.

Details

The functions blockStart and blockEnd create vectors of start and end values for equally sized 'timeDate' blocks. Note, the functions are event counters and not a time counter between measuring time intervals between start and end dates! For equally sized blocks in time one has before to align the time stamps in equal time differences.

Value

returns an object of class "timeDate".

Examples

  
## timeSequence
   # 360 Days Series:
   tS <- timeSequence(length.out = 360)
    
## blockStart | blockEnd -
   Start <- blockStart(tS, 30)
   End <- blockEnd(tS, 30)
   Start
   End
   End-Start

[Package timeDate version 3043.102 Index]