kurtosis {timeDate}	R Documentation

Kurtosis

Description

Functions to compute kurtosis.

Usage

kurtosis(x, ...)

## Default S3 method:
kurtosis(x, na.rm = FALSE, method = c("excess", "moment", "fisher"), ...)
## S3 method for class 'data.frame'
kurtosis(x, ...)
## S3 method for class 'POSIXct'
kurtosis(x, ...)
## S3 method for class 'POSIXlt'
kurtosis(x, ...)

Arguments

na.rm
a logical. Should missing values be removed?

method
a character string which specifies the method of computation. These are either "moment", "fisher", or "excess". If "excess" is selected, then the value of the kurtosis is computed by the "moment" method and a value of 3 will be subtracted. The "moment" method is based on the definitions of kurtosis for distributions; these forms should be used when resampling (bootstrap or jackknife). The "fisher" method correspond to the usual "unbiased" definition of sample variance, although in the case of kurtosis exact unbiasedness is not possible.

x
a numeric vector or object.

...
arguments to be passed.

Value

kurtosis

returns the value of the statistics, a numeric value. An attribute which reports the used method is added.

See Also

link{skewness}.

Examples

   
## mean -
## var -
   # Mean, Variance:
   r = rnorm(100)
   mean(r)
   var(r)
   
## kurtosis - 
   kurtosis(r)   

[Package timeDate version 3043.102 Index]