Sys.timeDate {timeDate}	R Documentation

System Time as 'timeDate' Object

Description

Returns system time as an object of class 'timeDate'.

Usage

Sys.timeDate(FinCenter = "")

Arguments

FinCenter
a character with the the location of the financial center named as "continent/city".

Value

returns the system time as class "timeDate" object.

Examples

## Sys.time -

   # direct
   Sys.timeDate()  
   
   # transformed from "POSIX(c)t"         
   timeDate(Sys.time()) 
 
   # Local Time in Zurich
   timeDate(Sys.time(), FinCenter = "Zurich")  

[Package timeDate version 3043.102 Index]