finCenter {timeDate}	R Documentation

Financial Center of a timeDate object

Description

Print or assign new financial center to a timeDate object.

Usage

## S4 method for signature 'timeDate'
finCenter(x)
## S4 replacement method for signature 'timeDate'
finCenter(x) <- value

Arguments

x
a timeSeries object.

value
a character with the the location of the financial center named as "continent/city".

See Also

listFinCenter

Examples

date <- timeDate("2008-01-01")
finCenter(date) <- "GMT"
date

finCenter(date) <- "Zurich"
date

[Package timeDate version 3043.102 Index]