List of Financial Centers

Lists supported financial centers.

Usage

listFinCenter(pattern = ".*")

Arguments

pattern

a pattern character string as required by the grep function. The default, ".*", gives all supported financial centers

Value

a character vector listing the financial centers whose names match

pattern.

See also

Examples

## myFinCenter - the global setting currently used:
   getRmetricsOptions("myFinCenter")
#> myFinCenter 
#>       "GMT" 

## Other Financial Centers:
   listFinCenter("Asia/")
#>  [1] "Asia/Aden"          "Asia/Almaty"        "Asia/Amman"        
#>  [4] "Asia/Anadyr"        "Asia/Aqtau"         "Asia/Aqtobe"       
#>  [7] "Asia/Ashgabat"      "Asia/Baghdad"       "Asia/Bahrain"      
#> [10] "Asia/Baku"          "Asia/Bangkok"       "Asia/Beirut"       
#> [13] "Asia/Bishkek"       "Asia/Brunei"        "Asia/Calcutta"     
#> [16] "Asia/Choibalsan"    "Asia/Chongqing"     "Asia/Colombo"      
#> [19] "Asia/Damascus"      "Asia/Dhaka"         "Asia/Dili"         
#> [22] "Asia/Dubai"         "Asia/Dushanbe"      "Asia/Gaza"         
#> [25] "Asia/Harbin"        "Asia/Hong_Kong"     "Asia/Hovd"         
#> [28] "Asia/Irkutsk"       "Asia/Jakarta"       "Asia/Jayapura"     
#> [31] "Asia/Jerusalem"     "Asia/Kabul"         "Asia/Kamchatka"    
#> [34] "Asia/Karachi"       "Asia/Kashgar"       "Asia/Katmandu"     
#> [37] "Asia/Krasnoyarsk"   "Asia/Kuala_Lumpur"  "Asia/Kuching"      
#> [40] "Asia/Kuwait"        "Asia/Macau"         "Asia/Magadan"      
#> [43] "Asia/Makassar"      "Asia/Manila"        "Asia/Muscat"       
#> [46] "Asia/Nicosia"       "Asia/Novosibirsk"   "Asia/Omsk"         
#> [49] "Asia/Oral"          "Asia/Phnom_Penh"    "Asia/Pontianak"    
#> [52] "Asia/Pyongyang"     "Asia/Qatar"         "Asia/Qyzylorda"    
#> [55] "Asia/Rangoon"       "Asia/Riyadh"        "Asia/Saigon"       
#> [58] "Asia/Sakhalin"      "Asia/Samarkand"     "Asia/Seoul"        
#> [61] "Asia/Shanghai"      "Asia/Singapore"     "Asia/Taipei"       
#> [64] "Asia/Tashkent"      "Asia/Tbilisi"       "Asia/Tehran"       
#> [67] "Asia/Thimphu"       "Asia/Tokyo"         "Asia/Ulaanbaatar"  
#> [70] "Asia/Urumqi"        "Asia/Vientiane"     "Asia/Vladivostok"  
#> [73] "Asia/Yakutsk"       "Asia/Yekaterinburg" "Asia/Yerevan"      
   listFinCenter("^A")    # all beginning with "A"
#>   [1] "Africa/Abidjan"                 "Africa/Accra"                  
#>   [3] "Africa/Addis_Ababa"             "Africa/Algiers"                
#>   [5] "Africa/Asmara"                  "Africa/Bamako"                 
#>   [7] "Africa/Bangui"                  "Africa/Banjul"                 
#>   [9] "Africa/Bissau"                  "Africa/Blantyre"               
#>  [11] "Africa/Brazzaville"             "Africa/Bujumbura"              
#>  [13] "Africa/Cairo"                   "Africa/Casablanca"             
#>  [15] "Africa/Ceuta"                   "Africa/Conakry"                
#>  [17] "Africa/Dakar"                   "Africa/Dar_es_Salaam"          
#>  [19] "Africa/Djibouti"                "Africa/Douala"                 
#>  [21] "Africa/El_Aaiun"                "Africa/Freetown"               
#>  [23] "Africa/Gaborone"                "Africa/Harare"                 
#>  [25] "Africa/Johannesburg"            "Africa/Kampala"                
#>  [27] "Africa/Khartoum"                "Africa/Kigali"                 
#>  [29] "Africa/Kinshasa"                "Africa/Lagos"                  
#>  [31] "Africa/Libreville"              "Africa/Lome"                   
#>  [33] "Africa/Luanda"                  "Africa/Lubumbashi"             
#>  [35] "Africa/Lusaka"                  "Africa/Malabo"                 
#>  [37] "Africa/Maputo"                  "Africa/Maseru"                 
#>  [39] "Africa/Mbabane"                 "Africa/Mogadishu"              
#>  [41] "Africa/Monrovia"                "Africa/Nairobi"                
#>  [43] "Africa/Ndjamena"                "Africa/Niamey"                 
#>  [45] "Africa/Nouakchott"              "Africa/Ouagadougou"            
#>  [47] "Africa/Porto-Novo"              "Africa/Sao_Tome"               
#>  [49] "Africa/Tripoli"                 "Africa/Tunis"                  
#>  [51] "Africa/Windhoek"                "America/Adak"                  
#>  [53] "America/Anchorage"              "America/Anguilla"              
#>  [55] "America/Antigua"                "America/Araguaina"             
#>  [57] "America/Argentina/Buenos_Aires" "America/Argentina/Catamarca"   
#>  [59] "America/Argentina/Cordoba"      "America/Argentina/Jujuy"       
#>  [61] "America/Argentina/La_Rioja"     "America/Argentina/Mendoza"     
#>  [63] "America/Argentina/Rio_Gallegos" "America/Argentina/San_Juan"    
#>  [65] "America/Argentina/Tucuman"      "America/Argentina/Ushuaia"     
#>  [67] "America/Aruba"                  "America/Asuncion"              
#>  [69] "America/Atikokan"               "America/Bahia"                 
#>  [71] "America/Barbados"               "America/Belem"                 
#>  [73] "America/Belize"                 "America/Blanc-Sablon"          
#>  [75] "America/Boa_Vista"              "America/Bogota"                
#>  [77] "America/Boise"                  "America/Cambridge_Bay"         
#>  [79] "America/Campo_Grande"           "America/Cancun"                
#>  [81] "America/Caracas"                "America/Cayenne"               
#>  [83] "America/Cayman"                 "America/Chicago"               
#>  [85] "America/Chihuahua"              "America/Costa_Rica"            
#>  [87] "America/Cuiaba"                 "America/Curacao"               
#>  [89] "America/Danmarkshavn"           "America/Dawson"                
#>  [91] "America/Dawson_Creek"           "America/Denver"                
#>  [93] "America/Detroit"                "America/Dominica"              
#>  [95] "America/Edmonton"               "America/Eirunepe"              
#>  [97] "America/El_Salvador"            "America/Fortaleza"             
#>  [99] "America/Glace_Bay"              "America/Godthab"               
#> [101] "America/Goose_Bay"              "America/Grand_Turk"            
#> [103] "America/Grenada"                "America/Guadeloupe"            
#> [105] "America/Guatemala"              "America/Guayaquil"             
#> [107] "America/Guyana"                 "America/Halifax"               
#> [109] "America/Havana"                 "America/Hermosillo"            
#> [111] "America/Indiana/Indianapolis"   "America/Indiana/Knox"          
#> [113] "America/Indiana/Marengo"        "America/Indiana/Petersburg"    
#> [115] "America/Indiana/Tell_City"      "America/Indiana/Vevay"         
#> [117] "America/Indiana/Vincennes"      "America/Indiana/Winamac"       
#> [119] "America/Inuvik"                 "America/Iqaluit"               
#> [121] "America/Jamaica"                "America/Juneau"                
#> [123] "America/Kentucky/Louisville"    "America/Kentucky/Monticello"   
#> [125] "America/La_Paz"                 "America/Lima"                  
#> [127] "America/Los_Angeles"            "America/Maceio"                
#> [129] "America/Managua"                "America/Manaus"                
#> [131] "America/Marigot"                "America/Martinique"            
#> [133] "America/Mazatlan"               "America/Menominee"             
#> [135] "America/Merida"                 "America/Mexico_City"           
#> [137] "America/Miquelon"               "America/Moncton"               
#> [139] "America/Monterrey"              "America/Montevideo"            
#> [141] "America/Montreal"               "America/Montserrat"            
#> [143] "America/Nassau"                 "America/New_York"              
#> [145] "America/Nipigon"                "America/Nome"                  
#> [147] "America/Noronha"                "America/North_Dakota/Center"   
#> [149] "America/North_Dakota/New_Salem" "America/Panama"                
#> [151] "America/Pangnirtung"            "America/Paramaribo"            
#> [153] "America/Phoenix"                "America/Port-au-Prince"        
#> [155] "America/Port_of_Spain"          "America/Porto_Velho"           
#> [157] "America/Puerto_Rico"            "America/Rainy_River"           
#> [159] "America/Rankin_Inlet"           "America/Recife"                
#> [161] "America/Regina"                 "America/Resolute"              
#> [163] "America/Rio_Branco"             "America/Santiago"              
#> [165] "America/Santo_Domingo"          "America/Sao_Paulo"             
#> [167] "America/Scoresbysund"           "America/Shiprock"              
#> [169] "America/St_Barthelemy"          "America/St_Johns"              
#> [171] "America/St_Kitts"               "America/St_Lucia"              
#> [173] "America/St_Thomas"              "America/St_Vincent"            
#> [175] "America/Swift_Current"          "America/Tegucigalpa"           
#> [177] "America/Thule"                  "America/Thunder_Bay"           
#> [179] "America/Tijuana"                "America/Toronto"               
#> [181] "America/Tortola"                "America/Vancouver"             
#> [183] "America/Whitehorse"             "America/Winnipeg"              
#> [185] "America/Yakutat"                "America/Yellowknife"           
#> [187] "Antarctica/Casey"               "Antarctica/Davis"              
#> [189] "Antarctica/DumontDUrville"      "Antarctica/Mawson"             
#> [191] "Antarctica/McMurdo"             "Antarctica/Palmer"             
#> [193] "Antarctica/Rothera"             "Antarctica/South_Pole"         
#> [195] "Antarctica/Syowa"               "Antarctica/Vostok"             
#> [197] "Arctic/Longyearbyen"            "Asia/Aden"                     
#> [199] "Asia/Almaty"                    "Asia/Amman"                    
#> [201] "Asia/Anadyr"                    "Asia/Aqtau"                    
#> [203] "Asia/Aqtobe"                    "Asia/Ashgabat"                 
#> [205] "Asia/Baghdad"                   "Asia/Bahrain"                  
#> [207] "Asia/Baku"                      "Asia/Bangkok"                  
#> [209] "Asia/Beirut"                    "Asia/Bishkek"                  
#> [211] "Asia/Brunei"                    "Asia/Calcutta"                 
#> [213] "Asia/Choibalsan"                "Asia/Chongqing"                
#> [215] "Asia/Colombo"                   "Asia/Damascus"                 
#> [217] "Asia/Dhaka"                     "Asia/Dili"                     
#> [219] "Asia/Dubai"                     "Asia/Dushanbe"                 
#> [221] "Asia/Gaza"                      "Asia/Harbin"                   
#> [223] "Asia/Hong_Kong"                 "Asia/Hovd"                     
#> [225] "Asia/Irkutsk"                   "Asia/Jakarta"                  
#> [227] "Asia/Jayapura"                  "Asia/Jerusalem"                
#> [229] "Asia/Kabul"                     "Asia/Kamchatka"                
#> [231] "Asia/Karachi"                   "Asia/Kashgar"                  
#> [233] "Asia/Katmandu"                  "Asia/Krasnoyarsk"              
#> [235] "Asia/Kuala_Lumpur"              "Asia/Kuching"                  
#> [237] "Asia/Kuwait"                    "Asia/Macau"                    
#> [239] "Asia/Magadan"                   "Asia/Makassar"                 
#> [241] "Asia/Manila"                    "Asia/Muscat"                   
#> [243] "Asia/Nicosia"                   "Asia/Novosibirsk"              
#> [245] "Asia/Omsk"                      "Asia/Oral"                     
#> [247] "Asia/Phnom_Penh"                "Asia/Pontianak"                
#> [249] "Asia/Pyongyang"                 "Asia/Qatar"                    
#> [251] "Asia/Qyzylorda"                 "Asia/Rangoon"                  
#> [253] "Asia/Riyadh"                    "Asia/Saigon"                   
#> [255] "Asia/Sakhalin"                  "Asia/Samarkand"                
#> [257] "Asia/Seoul"                     "Asia/Shanghai"                 
#> [259] "Asia/Singapore"                 "Asia/Taipei"                   
#> [261] "Asia/Tashkent"                  "Asia/Tbilisi"                  
#> [263] "Asia/Tehran"                    "Asia/Thimphu"                  
#> [265] "Asia/Tokyo"                     "Asia/Ulaanbaatar"              
#> [267] "Asia/Urumqi"                    "Asia/Vientiane"                
#> [269] "Asia/Vladivostok"               "Asia/Yakutsk"                  
#> [271] "Asia/Yekaterinburg"             "Asia/Yerevan"                  
#> [273] "Atlantic/Azores"                "Atlantic/Bermuda"              
#> [275] "Atlantic/Canary"                "Atlantic/Cape_Verde"           
#> [277] "Atlantic/Faroe"                 "Atlantic/Madeira"              
#> [279] "Atlantic/Reykjavik"             "Atlantic/South_Georgia"        
#> [281] "Atlantic/St_Helena"             "Atlantic/Stanley"              
#> [283] "Australia/Adelaide"             "Australia/Brisbane"            
#> [285] "Australia/Broken_Hill"          "Australia/Currie"              
#> [287] "Australia/Darwin"               "Australia/Eucla"               
#> [289] "Australia/Hobart"               "Australia/Lindeman"            
#> [291] "Australia/Lord_Howe"            "Australia/Melbourne"           
#> [293] "Australia/Perth"                "Australia/Sydney"              
   listFinCenter("^[^A]") # all *not* beginning with "A"
#>   [1] "Europe/Amsterdam"     "Europe/Andorra"       "Europe/Athens"       
#>   [4] "Europe/Belgrade"      "Europe/Berlin"        "Europe/Bratislava"   
#>   [7] "Europe/Brussels"      "Europe/Bucharest"     "Europe/Budapest"     
#>  [10] "Europe/Chisinau"      "Europe/Copenhagen"    "Europe/Dublin"       
#>  [13] "Europe/Gibraltar"     "Europe/Guernsey"      "Europe/Helsinki"     
#>  [16] "Europe/Isle_of_Man"   "Europe/Istanbul"      "Europe/Jersey"       
#>  [19] "Europe/Kaliningrad"   "Europe/Kiev"          "Europe/Lisbon"       
#>  [22] "Europe/Ljubljana"     "Europe/London"        "Europe/Luxembourg"   
#>  [25] "Europe/Madrid"        "Europe/Malta"         "Europe/Mariehamn"    
#>  [28] "Europe/Minsk"         "Europe/Monaco"        "Europe/Moscow"       
#>  [31] "Europe/Oslo"          "Europe/Paris"         "Europe/Podgorica"    
#>  [34] "Europe/Prague"        "Europe/Riga"          "Europe/Rome"         
#>  [37] "Europe/Samara"        "Europe/San_Marino"    "Europe/Sarajevo"     
#>  [40] "Europe/Simferopol"    "Europe/Skopje"        "Europe/Sofia"        
#>  [43] "Europe/Stockholm"     "Europe/Tallinn"       "Europe/Tirane"       
#>  [46] "Europe/Uzhgorod"      "Europe/Vaduz"         "Europe/Vatican"      
#>  [49] "Europe/Vienna"        "Europe/Vilnius"       "Europe/Volgograd"    
#>  [52] "Europe/Warsaw"        "Europe/Zagreb"        "Europe/Zaporozhye"   
#>  [55] "Europe/Zurich"        "Indian/Antananarivo"  "Indian/Chagos"       
#>  [58] "Indian/Christmas"     "Indian/Cocos"         "Indian/Comoro"       
#>  [61] "Indian/Kerguelen"     "Indian/Mahe"          "Indian/Maldives"     
#>  [64] "Indian/Mauritius"     "Indian/Mayotte"       "Indian/Reunion"      
#>  [67] "Pacific/Apia"         "Pacific/Auckland"     "Pacific/Chatham"     
#>  [70] "Pacific/Efate"        "Pacific/Enderbury"    "Pacific/Fakaofo"     
#>  [73] "Pacific/Fiji"         "Pacific/Funafuti"     "Pacific/Galapagos"   
#>  [76] "Pacific/Gambier"      "Pacific/Guadalcanal"  "Pacific/Guam"        
#>  [79] "Pacific/Honolulu"     "Pacific/Johnston"     "Pacific/Kiritimati"  
#>  [82] "Pacific/Kosrae"       "Pacific/Kwajalein"    "Pacific/Majuro"      
#>  [85] "Pacific/Marquesas"    "Pacific/Midway"       "Pacific/Nauru"       
#>  [88] "Pacific/Niue"         "Pacific/Norfolk"      "Pacific/Noumea"      
#>  [91] "Pacific/Pago_Pago"    "Pacific/Palau"        "Pacific/Pitcairn"    
#>  [94] "Pacific/Ponape"       "Pacific/Port_Moresby" "Pacific/Rarotonga"   
#>  [97] "Pacific/Saipan"       "Pacific/Tahiti"       "Pacific/Tarawa"      
#> [100] "Pacific/Tongatapu"    "Pacific/Truk"         "Pacific/Wake"        
#> [103] "Pacific/Wallis"      
   listFinCenter(".*/L")  # cities with L*
#>  [1] "Africa/Lagos"                "Africa/Libreville"          
#>  [3] "Africa/Lome"                 "Africa/Luanda"              
#>  [5] "Africa/Lubumbashi"           "Africa/Lusaka"              
#>  [7] "America/Argentina/La_Rioja"  "America/Kentucky/Louisville"
#>  [9] "America/La_Paz"              "America/Lima"               
#> [11] "America/Los_Angeles"         "Arctic/Longyearbyen"        
#> [13] "Australia/Lindeman"          "Australia/Lord_Howe"        
#> [15] "Europe/Lisbon"               "Europe/Ljubljana"           
#> [17] "Europe/London"               "Europe/Luxembourg"          

   stopifnot(identical(sort(listFinCenter()), ## 'A' and 'not A' == everything:
        sort(union(listFinCenter("^A"),
            listFinCenter("^[^A]")))))