listFinCenter {timeDate}	R Documentation

List of Financial Centers

Description

Lists supported financial centers.

Usage

listFinCenter(pattern = ".*")

Arguments

pattern
a pattern character string as required by the grep function.

Details

The function rulesFinCenter, lists the daylight saving rules for a selected financial center.

There is no dependency on the POSIX implementation of your operating system because all time zone and day light saving time information is stored locally in ASCII files.

Value

returns a list of supported financial centers.

Examples

## myFinCenter - the global setting currently used:
   getRmetricsOptions("myFinCenter")

## Other Financial Centers:
   listFinCenter("Asia/")
   listFinCenter("^A")    # all beginning with "A"
   listFinCenter("^[^A]") # all *not* beginning with "A"
   listFinCenter(".*/L")  # cities with L*

   stopifnot(identical(sort(listFinCenter()), ## 'A' and 'not A' == everything:
        sort(union(listFinCenter("^A"),
            listFinCenter("^[^A]")))))

[Package timeDate version 3043.102 Index]