Financial Centers DST Rules

Returns DST rules for a financial center.

Usage

rulesFinCenter(FinCenter = "")

Arguments

FinCenter

a character string with the location of the financial center named as "continent/city".

Value

a list of time zones and DST rules available in the database.

Details

The function rulesFinCenter lists the daylight saving rules for a selected financial center.

There is no dependency on the POSIX implementation of your operating system because all time zone and day light saving time information is stored locally in ASCII files.

See also

Examples

## rulesFinCenter - 
   rulesFinCenter("Zurich")
#>                  Zurich offSet isdst TimeZone     numeric
#> 1   1901-12-14 20:45:52   3600     0      CET -2147397248
#> 2   1941-05-05 00:00:00   7200     1     CEST  -904435200
#> 3   1941-10-06 00:00:00   3600     0      CET  -891129600
#> 4   1942-05-04 00:00:00   7200     1     CEST  -872985600
#> 5   1942-10-05 00:00:00   3600     0      CET  -859680000
#> 6   1981-03-29 01:00:00   7200     1     CEST   354675600
#> 7   1981-09-27 01:00:00   3600     0      CET   370400400
#> 8   1982-03-28 01:00:00   7200     1     CEST   386125200
#> 9   1982-09-26 01:00:00   3600     0      CET   401850000
#> 10  1983-03-27 01:00:00   7200     1     CEST   417574800
#> 11  1983-09-25 01:00:00   3600     0      CET   433299600
#> 12  1984-03-25 01:00:00   7200     1     CEST   449024400
#> 13  1984-09-30 01:00:00   3600     0      CET   465354000
#> 14  1985-03-31 01:00:00   7200     1     CEST   481078800
#> 15  1985-09-29 01:00:00   3600     0      CET   496803600
#> 16  1986-03-30 01:00:00   7200     1     CEST   512528400
#> 17  1986-09-28 01:00:00   3600     0      CET   528253200
#> 18  1987-03-29 01:00:00   7200     1     CEST   543978000
#> 19  1987-09-27 01:00:00   3600     0      CET   559702800
#> 20  1988-03-27 01:00:00   7200     1     CEST   575427600
#> 21  1988-09-25 01:00:00   3600     0      CET   591152400
#> 22  1989-03-26 01:00:00   7200     1     CEST   606877200
#> 23  1989-09-24 01:00:00   3600     0      CET   622602000
#> 24  1990-03-25 01:00:00   7200     1     CEST   638326800
#> 25  1990-09-30 01:00:00   3600     0      CET   654656400
#> 26  1991-03-31 01:00:00   7200     1     CEST   670381200
#> 27  1991-09-29 01:00:00   3600     0      CET   686106000
#> 28  1992-03-29 01:00:00   7200     1     CEST   701830800
#> 29  1992-09-27 01:00:00   3600     0      CET   717555600
#> 30  1993-03-28 01:00:00   7200     1     CEST   733280400
#> 31  1993-09-26 01:00:00   3600     0      CET   749005200
#> 32  1994-03-27 01:00:00   7200     1     CEST   764730000
#> 33  1994-09-25 01:00:00   3600     0      CET   780454800
#> 34  1995-03-26 01:00:00   7200     1     CEST   796179600
#> 35  1995-09-24 01:00:00   3600     0      CET   811904400
#> 36  1996-03-31 01:00:00   7200     1     CEST   828234000
#> 37  1996-10-27 01:00:00   3600     0      CET   846378000
#> 38  1997-03-30 01:00:00   7200     1     CEST   859683600
#> 39  1997-10-26 01:00:00   3600     0      CET   877827600
#> 40  1998-03-29 01:00:00   7200     1     CEST   891133200
#> 41  1998-10-25 01:00:00   3600     0      CET   909277200
#> 42  1999-03-28 01:00:00   7200     1     CEST   922582800
#> 43  1999-10-31 01:00:00   3600     0      CET   941331600
#> 44  2000-03-26 01:00:00   7200     1     CEST   954032400
#> 45  2000-10-29 01:00:00   3600     0      CET   972781200
#> 46  2001-03-25 01:00:00   7200     1     CEST   985482000
#> 47  2001-10-28 01:00:00   3600     0      CET  1004230800
#> 48  2002-03-31 01:00:00   7200     1     CEST  1017536400
#> 49  2002-10-27 01:00:00   3600     0      CET  1035680400
#> 50  2003-03-30 01:00:00   7200     1     CEST  1048986000
#> 51  2003-10-26 01:00:00   3600     0      CET  1067130000
#> 52  2004-03-28 01:00:00   7200     1     CEST  1080435600
#> 53  2004-10-31 01:00:00   3600     0      CET  1099184400
#> 54  2005-03-27 01:00:00   7200     1     CEST  1111885200
#> 55  2005-10-30 01:00:00   3600     0      CET  1130634000
#> 56  2006-03-26 01:00:00   7200     1     CEST  1143334800
#> 57  2006-10-29 01:00:00   3600     0      CET  1162083600
#> 58  2007-03-25 01:00:00   7200     1     CEST  1174784400
#> 59  2007-10-28 01:00:00   3600     0      CET  1193533200
#> 60  2008-03-30 01:00:00   7200     1     CEST  1206838800
#> 61  2008-10-26 01:00:00   3600     0      CET  1224982800
#> 62  2009-03-29 01:00:00   7200     1     CEST  1238288400
#> 63  2009-10-25 01:00:00   3600     0      CET  1256432400
#> 64  2010-03-28 01:00:00   7200     1     CEST  1269738000
#> 65  2010-10-31 01:00:00   3600     0      CET  1288486800
#> 66  2011-03-27 01:00:00   7200     1     CEST  1301187600
#> 67  2011-10-30 01:00:00   3600     0      CET  1319936400
#> 68  2012-03-25 01:00:00   7200     1     CEST  1332637200
#> 69  2012-10-28 01:00:00   3600     0      CET  1351386000
#> 70  2013-03-31 01:00:00   7200     1     CEST  1364691600
#> 71  2013-10-27 01:00:00   3600     0      CET  1382835600
#> 72  2014-03-30 01:00:00   7200     1     CEST  1396141200
#> 73  2014-10-26 01:00:00   3600     0      CET  1414285200
#> 74  2015-03-29 01:00:00   7200     1     CEST  1427590800
#> 75  2015-10-25 01:00:00   3600     0      CET  1445734800
#> 76  2016-03-27 01:00:00   7200     1     CEST  1459040400
#> 77  2016-10-30 01:00:00   3600     0      CET  1477789200
#> 78  2017-03-26 01:00:00   7200     1     CEST  1490490000
#> 79  2017-10-29 01:00:00   3600     0      CET  1509238800
#> 80  2018-03-25 01:00:00   7200     1     CEST  1521939600
#> 81  2018-10-28 01:00:00   3600     0      CET  1540688400
#> 82  2019-03-31 01:00:00   7200     1     CEST  1553994000
#> 83  2019-10-27 01:00:00   3600     0      CET  1572138000
#> 84  2020-03-29 01:00:00   7200     1     CEST  1585443600
#> 85  2020-10-25 01:00:00   3600     0      CET  1603587600
#> 86  2021-03-28 01:00:00   7200     1     CEST  1616893200
#> 87  2021-10-31 01:00:00   3600     0      CET  1635642000
#> 88  2022-03-27 01:00:00   7200     1     CEST  1648342800
#> 89  2022-10-30 01:00:00   3600     0      CET  1667091600
#> 90  2023-03-26 01:00:00   7200     1     CEST  1679792400
#> 91  2023-10-29 01:00:00   3600     0      CET  1698541200
#> 92  2024-03-31 01:00:00   7200     1     CEST  1711846800
#> 93  2024-10-27 01:00:00   3600     0      CET  1729990800
#> 94  2025-03-30 01:00:00   7200     1     CEST  1743296400
#> 95  2025-10-26 01:00:00   3600     0      CET  1761440400
#> 96  2026-03-29 01:00:00   7200     1     CEST  1774746000
#> 97  2026-10-25 01:00:00   3600     0      CET  1792890000
#> 98  2027-03-28 01:00:00   7200     1     CEST  1806195600
#> 99  2027-10-31 01:00:00   3600     0      CET  1824944400
#> 100 2028-03-26 01:00:00   7200     1     CEST  1837645200
#> 101 2028-10-29 01:00:00   3600     0      CET  1856394000
#> 102 2029-03-25 01:00:00   7200     1     CEST  1869094800
#> 103 2029-10-28 01:00:00   3600     0      CET  1887843600
#> 104 2030-03-31 01:00:00   7200     1     CEST  1901149200
#> 105 2030-10-27 01:00:00   3600     0      CET  1919293200
#> 106 2031-03-30 01:00:00   7200     1     CEST  1932598800
#> 107 2031-10-26 01:00:00   3600     0      CET  1950742800
#> 108 2032-03-28 01:00:00   7200     1     CEST  1964048400
#> 109 2032-10-31 01:00:00   3600     0      CET  1982797200
#> 110 2033-03-27 01:00:00   7200     1     CEST  1995498000
#> 111 2033-10-30 01:00:00   3600     0      CET  2014246800
#> 112 2034-03-26 01:00:00   7200     1     CEST  2026947600
#> 113 2034-10-29 01:00:00   3600     0      CET  2045696400
#> 114 2035-03-25 01:00:00   7200     1     CEST  2058397200
#> 115 2035-10-28 01:00:00   3600     0      CET  2077146000
#> 116 2036-03-30 01:00:00   7200     1     CEST  2090451600
#> 117 2036-10-26 01:00:00   3600     0      CET  2108595600
#> 118 2037-03-29 01:00:00   7200     1     CEST  2121901200
#> 119 2037-10-25 01:00:00   3600     0      CET  2140045200