Financial centers DST rules

Returns DST rules for a financial center.

Usage

rulesFinCenter(FinCenter = "")

Arguments

FinCenter

a character string with the location of the financial center named as "continent/city".

Value

a list of time zones and DST rules available in the database

Details

The function rulesFinCenter lists the daylight saving rules for a selected financial center.

There is no dependency on the POSIX implementation of your operating system because timeDate comes with a database containing the necessary time zone and day light saving time information.

See also

Examples

## rulesFinCenter - 
   rulesFinCenter("Zurich")
#>                  Zurich offSet isdst TimeZone     numeric
#> 1   1853-07-15 23:25:52   1786     0      BMT -3675198848
#> 2   1894-05-31 23:30:14   3600     0      CET -2385246586
#> 3   1941-05-05 00:00:00   7200     1     CEST  -904435200
#> 4   1941-10-06 00:00:00   3600     0      CET  -891129600
#> 5   1942-05-04 00:00:00   7200     1     CEST  -872985600
#> 6   1942-10-05 00:00:00   3600     0      CET  -859680000
#> 7   1981-03-29 01:00:00   7200     1     CEST   354675600
#> 8   1981-09-27 01:00:00   3600     0      CET   370400400
#> 9   1982-03-28 01:00:00   7200     1     CEST   386125200
#> 10  1982-09-26 01:00:00   3600     0      CET   401850000
#> 11  1983-03-27 01:00:00   7200     1     CEST   417574800
#> 12  1983-09-25 01:00:00   3600     0      CET   433299600
#> 13  1984-03-25 01:00:00   7200     1     CEST   449024400
#> 14  1984-09-30 01:00:00   3600     0      CET   465354000
#> 15  1985-03-31 01:00:00   7200     1     CEST   481078800
#> 16  1985-09-29 01:00:00   3600     0      CET   496803600
#> 17  1986-03-30 01:00:00   7200     1     CEST   512528400
#> 18  1986-09-28 01:00:00   3600     0      CET   528253200
#> 19  1987-03-29 01:00:00   7200     1     CEST   543978000
#> 20  1987-09-27 01:00:00   3600     0      CET   559702800
#> 21  1988-03-27 01:00:00   7200     1     CEST   575427600
#> 22  1988-09-25 01:00:00   3600     0      CET   591152400
#> 23  1989-03-26 01:00:00   7200     1     CEST   606877200
#> 24  1989-09-24 01:00:00   3600     0      CET   622602000
#> 25  1990-03-25 01:00:00   7200     1     CEST   638326800
#> 26  1990-09-30 01:00:00   3600     0      CET   654656400
#> 27  1991-03-31 01:00:00   7200     1     CEST   670381200
#> 28  1991-09-29 01:00:00   3600     0      CET   686106000
#> 29  1992-03-29 01:00:00   7200     1     CEST   701830800
#> 30  1992-09-27 01:00:00   3600     0      CET   717555600
#> 31  1993-03-28 01:00:00   7200     1     CEST   733280400
#> 32  1993-09-26 01:00:00   3600     0      CET   749005200
#> 33  1994-03-27 01:00:00   7200     1     CEST   764730000
#> 34  1994-09-25 01:00:00   3600     0      CET   780454800
#> 35  1995-03-26 01:00:00   7200     1     CEST   796179600
#> 36  1995-09-24 01:00:00   3600     0      CET   811904400
#> 37  1996-03-31 01:00:00   7200     1     CEST   828234000
#> 38  1996-10-27 01:00:00   3600     0      CET   846378000
#> 39  1997-03-30 01:00:00   7200     1     CEST   859683600
#> 40  1997-10-26 01:00:00   3600     0      CET   877827600
#> 41  1998-03-29 01:00:00   7200     1     CEST   891133200
#> 42  1998-10-25 01:00:00   3600     0      CET   909277200
#> 43  1999-03-28 01:00:00   7200     1     CEST   922582800
#> 44  1999-10-31 01:00:00   3600     0      CET   941331600
#> 45  2000-03-26 01:00:00   7200     1     CEST   954032400
#> 46  2000-10-29 01:00:00   3600     0      CET   972781200
#> 47  2001-03-25 01:00:00   7200     1     CEST   985482000
#> 48  2001-10-28 01:00:00   3600     0      CET  1004230800
#> 49  2002-03-31 01:00:00   7200     1     CEST  1017536400
#> 50  2002-10-27 01:00:00   3600     0      CET  1035680400
#> 51  2003-03-30 01:00:00   7200     1     CEST  1048986000
#> 52  2003-10-26 01:00:00   3600     0      CET  1067130000
#> 53  2004-03-28 01:00:00   7200     1     CEST  1080435600
#> 54  2004-10-31 01:00:00   3600     0      CET  1099184400
#> 55  2005-03-27 01:00:00   7200     1     CEST  1111885200
#> 56  2005-10-30 01:00:00   3600     0      CET  1130634000
#> 57  2006-03-26 01:00:00   7200     1     CEST  1143334800
#> 58  2006-10-29 01:00:00   3600     0      CET  1162083600
#> 59  2007-03-25 01:00:00   7200     1     CEST  1174784400
#> 60  2007-10-28 01:00:00   3600     0      CET  1193533200
#> 61  2008-03-30 01:00:00   7200     1     CEST  1206838800
#> 62  2008-10-26 01:00:00   3600     0      CET  1224982800
#> 63  2009-03-29 01:00:00   7200     1     CEST  1238288400
#> 64  2009-10-25 01:00:00   3600     0      CET  1256432400
#> 65  2010-03-28 01:00:00   7200     1     CEST  1269738000
#> 66  2010-10-31 01:00:00   3600     0      CET  1288486800
#> 67  2011-03-27 01:00:00   7200     1     CEST  1301187600
#> 68  2011-10-30 01:00:00   3600     0      CET  1319936400
#> 69  2012-03-25 01:00:00   7200     1     CEST  1332637200
#> 70  2012-10-28 01:00:00   3600     0      CET  1351386000
#> 71  2013-03-31 01:00:00   7200     1     CEST  1364691600
#> 72  2013-10-27 01:00:00   3600     0      CET  1382835600
#> 73  2014-03-30 01:00:00   7200     1     CEST  1396141200
#> 74  2014-10-26 01:00:00   3600     0      CET  1414285200
#> 75  2015-03-29 01:00:00   7200     1     CEST  1427590800
#> 76  2015-10-25 01:00:00   3600     0      CET  1445734800
#> 77  2016-03-27 01:00:00   7200     1     CEST  1459040400
#> 78  2016-10-30 01:00:00   3600     0      CET  1477789200
#> 79  2017-03-26 01:00:00   7200     1     CEST  1490490000
#> 80  2017-10-29 01:00:00   3600     0      CET  1509238800
#> 81  2018-03-25 01:00:00   7200     1     CEST  1521939600
#> 82  2018-10-28 01:00:00   3600     0      CET  1540688400
#> 83  2019-03-31 01:00:00   7200     1     CEST  1553994000
#> 84  2019-10-27 01:00:00   3600     0      CET  1572138000
#> 85  2020-03-29 01:00:00   7200     1     CEST  1585443600
#> 86  2020-10-25 01:00:00   3600     0      CET  1603587600
#> 87  2021-03-28 01:00:00   7200     1     CEST  1616893200
#> 88  2021-10-31 01:00:00   3600     0      CET  1635642000
#> 89  2022-03-27 01:00:00   7200     1     CEST  1648342800
#> 90  2022-10-30 01:00:00   3600     0      CET  1667091600
#> 91  2023-03-26 01:00:00   7200     1     CEST  1679792400
#> 92  2023-10-29 01:00:00   3600     0      CET  1698541200
#> 93  2024-03-31 01:00:00   7200     1     CEST  1711846800
#> 94  2024-10-27 01:00:00   3600     0      CET  1729990800
#> 95  2025-03-30 01:00:00   7200     1     CEST  1743296400
#> 96  2025-10-26 01:00:00   3600     0      CET  1761440400
#> 97  2026-03-29 01:00:00   7200     1     CEST  1774746000
#> 98  2026-10-25 01:00:00   3600     0      CET  1792890000
#> 99  2027-03-28 01:00:00   7200     1     CEST  1806195600
#> 100 2027-10-31 01:00:00   3600     0      CET  1824944400
#> 101 2028-03-26 01:00:00   7200     1     CEST  1837645200
#> 102 2028-10-29 01:00:00   3600     0      CET  1856394000
#> 103 2029-03-25 01:00:00   7200     1     CEST  1869094800
#> 104 2029-10-28 01:00:00   3600     0      CET  1887843600
#> 105 2030-03-31 01:00:00   7200     1     CEST  1901149200
#> 106 2030-10-27 01:00:00   3600     0      CET  1919293200
#> 107 2031-03-30 01:00:00   7200     1     CEST  1932598800
#> 108 2031-10-26 01:00:00   3600     0      CET  1950742800
#> 109 2032-03-28 01:00:00   7200     1     CEST  1964048400
#> 110 2032-10-31 01:00:00   3600     0      CET  1982797200
#> 111 2033-03-27 01:00:00   7200     1     CEST  1995498000
#> 112 2033-10-30 01:00:00   3600     0      CET  2014246800
#> 113 2034-03-26 01:00:00   7200     1     CEST  2026947600
#> 114 2034-10-29 01:00:00   3600     0      CET  2045696400
#> 115 2035-03-25 01:00:00   7200     1     CEST  2058397200
#> 116 2035-10-28 01:00:00   3600     0      CET  2077146000
#> 117 2036-03-30 01:00:00   7200     1     CEST  2090451600
#> 118 2036-10-26 01:00:00   3600     0      CET  2108595600
#> 119 2037-03-29 01:00:00   7200     1     CEST  2121901200
#> 120 2037-10-25 01:00:00   3600     0      CET  2140045200
#> 121 2038-03-28 01:00:00   7200     1     CEST  2153350800
#> 122 2038-10-31 01:00:00   3600     0      CET  2172099600
#> 123 2039-03-27 01:00:00   7200     1     CEST  2184800400
#> 124 2039-10-30 01:00:00   3600     0      CET  2203549200
#> 125 2040-03-25 01:00:00   7200     1     CEST  2216250000
#> 126 2040-10-28 01:00:00   3600     0      CET  2234998800
#> 127 2041-03-31 01:00:00   7200     1     CEST  2248304400
#> 128 2041-10-27 01:00:00   3600     0      CET  2266448400
#> 129 2042-03-30 01:00:00   7200     1     CEST  2279754000
#> 130 2042-10-26 01:00:00   3600     0      CET  2297898000
#> 131 2043-03-29 01:00:00   7200     1     CEST  2311203600
#> 132 2043-10-25 01:00:00   3600     0      CET  2329347600
#> 133 2044-03-27 01:00:00   7200     1     CEST  2342653200
#> 134 2044-10-30 01:00:00   3600     0      CET  2361402000
#> 135 2045-03-26 01:00:00   7200     1     CEST  2374102800
#> 136 2045-10-29 01:00:00   3600     0      CET  2392851600
#> 137 2046-03-25 01:00:00   7200     1     CEST  2405552400
#> 138 2046-10-28 01:00:00   3600     0      CET  2424301200
#> 139 2047-03-31 01:00:00   7200     1     CEST  2437606800
#> 140 2047-10-27 01:00:00   3600     0      CET  2455750800
#> 141 2048-03-29 01:00:00   7200     1     CEST  2469056400
#> 142 2048-10-25 01:00:00   3600     0      CET  2487200400
#> 143 2049-03-28 01:00:00   7200     1     CEST  2500506000
#> 144 2049-10-31 01:00:00   3600     0      CET  2519254800
#> 145 2050-03-27 01:00:00   7200     1     CEST  2531955600
#> 146 2050-10-30 01:00:00   3600     0      CET  2550704400
#> 147 2051-03-26 01:00:00   7200     1     CEST  2563405200
#> 148 2051-10-29 01:00:00   3600     0      CET  2582154000
#> 149 2052-03-31 01:00:00   7200     1     CEST  2595459600
#> 150 2052-10-27 01:00:00   3600     0      CET  2613603600
#> 151 2053-03-30 01:00:00   7200     1     CEST  2626909200
#> 152 2053-10-26 01:00:00   3600     0      CET  2645053200
#> 153 2054-03-29 01:00:00   7200     1     CEST  2658358800
#> 154 2054-10-25 01:00:00   3600     0      CET  2676502800
#> 155 2055-03-28 01:00:00   7200     1     CEST  2689808400
#> 156 2055-10-31 01:00:00   3600     0      CET  2708557200
#> 157 2056-03-26 01:00:00   7200     1     CEST  2721258000
#> 158 2056-10-29 01:00:00   3600     0      CET  2740006800
#> 159 2057-03-25 01:00:00   7200     1     CEST  2752707600
#> 160 2057-10-28 01:00:00   3600     0      CET  2771456400
#> 161 2058-03-31 01:00:00   7200     1     CEST  2784762000
#> 162 2058-10-27 01:00:00   3600     0      CET  2802906000
#> 163 2059-03-30 01:00:00   7200     1     CEST  2816211600
#> 164 2059-10-26 01:00:00   3600     0      CET  2834355600
#> 165 2060-03-28 01:00:00   7200     1     CEST  2847661200
#> 166 2060-10-31 01:00:00   3600     0      CET  2866410000
#> 167 2061-03-27 01:00:00   7200     1     CEST  2879110800
#> 168 2061-10-30 01:00:00   3600     0      CET  2897859600
#> 169 2062-03-26 01:00:00   7200     1     CEST  2910560400
#> 170 2062-10-29 01:00:00   3600     0      CET  2929309200
#> 171 2063-03-25 01:00:00   7200     1     CEST  2942010000
#> 172 2063-10-28 01:00:00   3600     0      CET  2960758800
#> 173 2064-03-30 01:00:00   7200     1     CEST  2974064400
#> 174 2064-10-26 01:00:00   3600     0      CET  2992208400
#> 175 2065-03-29 01:00:00   7200     1     CEST  3005514000
#> 176 2065-10-25 01:00:00   3600     0      CET  3023658000
#> 177 2066-03-28 01:00:00   7200     1     CEST  3036963600
#> 178 2066-10-31 01:00:00   3600     0      CET  3055712400
#> 179 2067-03-27 01:00:00   7200     1     CEST  3068413200
#> 180 2067-10-30 01:00:00   3600     0      CET  3087162000
#> 181 2068-03-25 01:00:00   7200     1     CEST  3099862800
#> 182 2068-10-28 01:00:00   3600     0      CET  3118611600
#> 183 2069-03-31 01:00:00   7200     1     CEST  3131917200
#> 184 2069-10-27 01:00:00   3600     0      CET  3150061200
#> 185 2070-03-30 01:00:00   7200     1     CEST  3163366800
#> 186 2070-10-26 01:00:00   3600     0      CET  3181510800
#> 187 2071-03-29 01:00:00   7200     1     CEST  3194816400
#> 188 2071-10-25 01:00:00   3600     0      CET  3212960400
#> 189 2072-03-27 01:00:00   7200     1     CEST  3226266000
#> 190 2072-10-30 01:00:00   3600     0      CET  3245014800
#> 191 2073-03-26 01:00:00   7200     1     CEST  3257715600
#> 192 2073-10-29 01:00:00   3600     0      CET  3276464400
#> 193 2074-03-25 01:00:00   7200     1     CEST  3289165200
#> 194 2074-10-28 01:00:00   3600     0      CET  3307914000
#> 195 2075-03-31 01:00:00   7200     1     CEST  3321219600
#> 196 2075-10-27 01:00:00   3600     0      CET  3339363600
#> 197 2076-03-29 01:00:00   7200     1     CEST  3352669200
#> 198 2076-10-25 01:00:00   3600     0      CET  3370813200
#> 199 2077-03-28 01:00:00   7200     1     CEST  3384118800
#> 200 2077-10-31 01:00:00   3600     0      CET  3402867600
#> 201 2078-03-27 01:00:00   7200     1     CEST  3415568400
#> 202 2078-10-30 01:00:00   3600     0      CET  3434317200
#> 203 2079-03-26 01:00:00   7200     1     CEST  3447018000
#> 204 2079-10-29 01:00:00   3600     0      CET  3465766800
#> 205 2080-03-31 01:00:00   7200     1     CEST  3479072400
#> 206 2080-10-27 01:00:00   3600     0      CET  3497216400
#> 207 2081-03-30 01:00:00   7200     1     CEST  3510522000
#> 208 2081-10-26 01:00:00   3600     0      CET  3528666000
#> 209 2082-03-29 01:00:00   7200     1     CEST  3541971600
#> 210 2082-10-25 01:00:00   3600     0      CET  3560115600
#> 211 2083-03-28 01:00:00   7200     1     CEST  3573421200
#> 212 2083-10-31 01:00:00   3600     0      CET  3592170000
#> 213 2084-03-26 01:00:00   7200     1     CEST  3604870800
#> 214 2084-10-29 01:00:00   3600     0      CET  3623619600
#> 215 2085-03-25 01:00:00   7200     1     CEST  3636320400
#> 216 2085-10-28 01:00:00   3600     0      CET  3655069200
#> 217 2086-03-31 01:00:00   7200     1     CEST  3668374800
#> 218 2086-10-27 01:00:00   3600     0      CET  3686518800
#> 219 2087-03-30 01:00:00   7200     1     CEST  3699824400
#> 220 2087-10-26 01:00:00   3600     0      CET  3717968400
#> 221 2088-03-28 01:00:00   7200     1     CEST  3731274000
#> 222 2088-10-31 01:00:00   3600     0      CET  3750022800
#> 223 2089-03-27 01:00:00   7200     1     CEST  3762723600
#> 224 2089-10-30 01:00:00   3600     0      CET  3781472400
#> 225 2090-03-26 01:00:00   7200     1     CEST  3794173200
#> 226 2090-10-29 01:00:00   3600     0      CET  3812922000
#> 227 2091-03-25 01:00:00   7200     1     CEST  3825622800
#> 228 2091-10-28 01:00:00   3600     0      CET  3844371600
#> 229 2092-03-30 01:00:00   7200     1     CEST  3857677200
#> 230 2092-10-26 01:00:00   3600     0      CET  3875821200
#> 231 2093-03-29 01:00:00   7200     1     CEST  3889126800
#> 232 2093-10-25 01:00:00   3600     0      CET  3907270800
#> 233 2094-03-28 01:00:00   7200     1     CEST  3920576400
#> 234 2094-10-31 01:00:00   3600     0      CET  3939325200
#> 235 2095-03-27 01:00:00   7200     1     CEST  3952026000
#> 236 2095-10-30 01:00:00   3600     0      CET  3970774800
#> 237 2096-03-25 01:00:00   7200     1     CEST  3983475600
#> 238 2096-10-28 01:00:00   3600     0      CET  4002224400
#> 239 2097-03-31 01:00:00   7200     1     CEST  4015530000
#> 240 2097-10-27 01:00:00   3600     0      CET  4033674000
#> 241 2098-03-30 01:00:00   7200     1     CEST  4046979600
#> 242 2098-10-26 01:00:00   3600     0      CET  4065123600
#> 243 2099-03-29 01:00:00   7200     1     CEST  4078429200
#> 244 2099-10-25 01:00:00   3600     0      CET  4096573200