rulesFinCenter {timeDate}	R Documentation

Financial Centers DST Rules

Description

Returns DST rules for a financial center.

Usage

rulesFinCenter(FinCenter = "")

Arguments

FinCenter
a character with the the location of the financial center named as "continent/city".

Details

The function rulesFinCenter, lists the daylight saving rules for a selected financial center.

There is no dependency on the POSIX implementation of your operating system because all time zone and day light saving time information is stored locally in ASCII files.

Value

returns a list of time zones and DST rules available in the database.

Examples

## rulesFinCenter - 
   rulesFinCenter("Zurich")

[Package timeDate version 3043.102 Index]