whichFormat {timeDate}	R Documentation

Format Recognation

Description

Trys to recognize the date/time format.

Usage

whichFormat(charvec, silent = FALSE)  

Arguments

charvec
a character string or vector of dates and times.

silent
a logical flag. Should a warning be printed if the format cannot be recognized?

Value

returns a format string.

Examples

## midnightStandard -  
   whichFormat("2007-12-31 24:00")

[Package timeDate version 3043.102 Index]